### 2025-09-30 #265 @Nischalcs50
nsEVDx a Python Package for modeling non-stationary extreme value distributions - Submission for Review
Submitting Author: (@Nischalcs50)
Package Name: nsEVDx
One-Line Description of Package: This package implements both bayesian sampling techniques and frequentist methods to estimate the parameters of non-stationary extreme value distributions
Repository Link:  https://github.com/Nischalcs50/nsEVDx
Version submitted: v0.1.1
Version accepted: v0.2.3
Presubmission issue: https://github.com/pyOpenSci/software-submission/issues/264#issuecomment-3332066925

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

- nsEVDx is a Python library for estimating the parameters of Generalized Extreme Value (GEV) and Generalized Pareto Distributions (GPD), collectively referred to as extreme value distributions (EVDs), under both stationary and non-stationary assumptions, using frequentist and Bayesian methods. Designed for hydrologists, climate scientists, and engineers, especially those working on extreme rainfall or flood frequency analysis, it supports time-varying covariates, MCMC samplings (Metropolis hasting-Randomwalk, Adjusted Langevin Algorithm, Hamiltonian Monte Carlo) and essential model diagnostics. Although developed for environmental extremes, its features are broadly applicable to financial risk modeling and other domains concerned with rare, high-impact events.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.
