### 2026-02-01 #265 @crhea93
- [x] **Installation** The package can be installed from a community repository such as PyPI (preferred), and/or a community channel on conda (for example, conda-forge, bioconda).
- [x] **Documentation** The package has sufficient online documentation to allow us to evaluate the package's function and scope *without installing the package*. This includes:
  - [x] Short, quickstart tutorials that help a user understand how to use the package and what it can do for them.
  - [x] API documentation (documentation for your code's functions, classes, methods, and attributes): this includes clearly written docstrings with variables defined using a standard docstring format.
  - [x] **README** The package has a `README.md` file with a clear explanation of what the package does, instructions on how to install it, and a link to development instructions.
- [ ] The package authors have disclosed the use of Generative AI tools in the development and/or maintenance of their package.

**Optional but recommended:**

- [ ] **Linting** The package uses automated linting (e.g., [ruff](https://docs.astral.sh/ruff/)) to enforce code style and quality standards, ideally integrated with CI.
- [ ] **Type checking** The package includes type hints and uses a type checker (e.g., [mypy](https://mypy-lang.org/)) to catch type-related issues, ideally integrated with CI.
- [ ] **Modern tooling** The package uses modern Python development tools (e.g., [uv](https://docs.astral.sh/uv/) for dependency management) and [Hatch](https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-build-tools.html#about-hatch) for building publishing and running development tasks to streamline development workflows. See the pyOpenSci packaging guide for more on [modern packaging tools](https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-build-tools.html).

**Optional:** Let projects know that it's a great idea for projects to have a .github repository for the project organization where they can host a commonly used LICENSE, Code of Conduct, and even a YAML file with label definitions. These items will then be automatically applied to every repository in the organization to ensure consistency (but can be customized within repos too). The [SunPy project](https://github.com/sunpy/.github/) has a great example of this.

---
A big thank you in advance to authors for setting aside five to ten minutes to do this. It truly helps our organization! :raised_hands:
---

### Major Comments
- The documentation looks to be built by hand. While the author has certainly built a very nice documentation set, they may find this difficult to maintain. I'd strongly suggest switching to a more robust and automated method such as [Sphinx](https://www.sphinx-doc.org/en/master/usage/installation.html) for the documentation (including API documentation).
- Issue https://github.com/Nischalcs50/nsEVDx/issues/2
- There does not appear to be any instructions on how to run the unit tests in the `tests` directory. I'd suggest setting up a `pytest.ini` file (or a non pytest equivalent if you prefer) for ease of unit testing. Additionally, we request that all authors include continuous integration via GitHub actions for unit tests.

- It would be a nice addition if the code snippets in the documentation could be copied.
- I strongly suggest adopting an automated linter. This can be done nicely in either a pre commit hook (i.e. https://pre-commit.com/) with ruff (https://docs.astral.sh/ruff/) or a GitHub Action.
