### 2026-02-20 #265 @JuliMillan
## Editor response to review:

---

## Editor comments

Hi :wave: Thank you @akritaag and @adityapt for volunterring to review nsEVDx!

2. Here is the **[review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template)** that you will need to fill out and submit here as a comment, once your review is complete.

Please get in touch with any questions or concerns! Your review is due: March 17 2026
---

Reviewers: @akritaag @adityapt
Due date: March 17
