### 2026-02-23 aa3d024
Fix link to examples in README (fixed #3)
Thank you for point out the issue
