### 2026-03-08 #265 @adityapt
Hi @JuliMillan and @Nischalcs50  - below is my review. Overall, it's good, but there are some gaps which need to be addressed.

Creating my review template. Will keep updating it.

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [ ] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

2.5 hours

---

1) The tests have passed in my local: 17 passed, 16 warnings in 6.69s
2) There is no badge for CI/CD in the ReadMe.MD file, and the test coverage seems to be minimal.

high-level overview:

Test coverage: ~74% overall. init and _version are fully covered; these have evd_model.py ~68%, utils.py ~80% coverage.
Gaps in Testing: Some of the below don't have tests setup.
error paths (invalid prior/bounds), log_prior edge cases, numerical_grad_log_posterior, hamiltonian, MALA/HMC internals, frequentist_nsEVD retries/fallback, and gelman_rubin.

Tests: 17 tests, all are passing. One bug: duplicate name test_MH_Mala (second is actually HMC), so the HMC test overwrites the MALA one. However, there are no tests for log_prior, suggest_bounds/suggest_priors edge cases, or ns_EVDrvs size mismatch.

PEP 8 / style: No config (no flake8, ruff, or pyproject style section).
Minor issues: inconsistent spacing (e.g. [0.0,0,1.0, 0.1]), mixed quote style, some long lines and trailing whitespace, print() used for warnings/optimization messages instead of logging, and type hints only in a few places (e.g. utils.py has a typo in a type hint: np.ndarray]).

Package / repo: setup.py and pyproject.toml disagree on deps (e.g. pandas/tqdm only in setup.py). No pytest-cov in dev deps. No no pre-commit or CI config.

I see quite some CI gaps:

1) Single Python version — only 3.9; please add 3.10 and 3.11 (and 3.12 if you support it) for compatibility.
2) No lint/format job — no flake8/ruff/isort/black in CI.
3) Fragile install — CI uses requirements.txt plus .[dev]; pyproject.toml and setup.py/setup.cfg list different deps (

Also:

No release workflow (e.g. trigger on tag, build wheel/sdist, publish to PyPI).
No CHANGELOG.md.
No SECURITY.md.
CONTRIBUTING.MD has a broken link: [MIT] -> [MIT] lisense (there is a typo in license) and no URL is present.
