### 2026-03-09 #265 @akritaag
Hi @Nischalcs50 and @JuliMillan I just finished reviewing the package -

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

- [x] **Performance:** Any performance claims of the software been confirmed. (**No claims made**)

---

Can you review all the items above that were not checked, summarized the major action items here:

1. > **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
Quickstart has ellipses and no actual examples. The documentation section following the Quick-start has user examples and this guide has quickstart examples too: https://nischalcs50.github.io/nsEVDx/

Perhaps we can copy the actual quickstart over to the package? I have checked-in with an experienced reviewer for the strictness here.

2. The readme does not have a badge for continuous integration/test coverage.
3. Vignette coverage/demonstrations incomplete: the README links to a handful of example notebooks but doesn’t enumerate all vignettes or offer a runnable usage demonstration, so this box remains unchecked.
4. No comparison to related packages or positioning within the broader ecosystem is provided.
5. `paper.md` is missing authors and references, I see authors.md has that information so you want to port that into the paper.md file per the JOSS guidelines.
6. Is this package also submitted to JOSS?
7. The package code doesn't consistently follow the PEP 8 guidelines. Can you fix this and upload the results for the following command -
.venv_install/bin/flake8 nsEVDx tests
  └ nsEVDx/evd_model.py:2:1: F401 'matplotlib.pyplot as plt' imported but unused
    nsEVDx/evd_model.py:3:1: F401 'scipy.stats.chi2' imported but unused
    … +356 lines
    tests/test_utils.py:56:1: E302 expected 2 blank lines, found 1
    tests/test_utils.py:62:1: E302 expected 2 blank lines, found 1

```
8. Are you able to add test cases for these scenarios -
```
1. Add cases for GPD modelling and non-stationary scale/shape paths.
  2. Exercise diagnostics/utility functions where coverage is low (e.g. sections around nsEVDx/utils.py:430 and nsEVDx/evd_model.py:724).
  3. Consider edge-condition tests (small samples, extreme covariates) to broaden input coverage.
```

Let me know once you've reviewed/addressed the comments, I can check again
