### 2026-03-09 d4b65f6
added new development badges to readme.md and arranged them into tabular format
