### 2026-03-09 8aa7e67
added quick start example notebook and also in the readme.md
