### 2026-03-09 aac7ad8
removed old docs for web documentation and created new using sphinx
