+++ docs/source/nsEVDx.rst
@@ -1,0 +1,29 @@
+nsEVDx package
+==============
+
+Submodules
+----------
+
+nsEVDx.evd\_model module
+------------------------
+
+.. automodule:: nsEVDx.evd_model
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+nsEVDx.utils module
+-------------------
+
+.. automodule:: nsEVDx.utils
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: nsEVDx
+   :members:
+   :show-inheritance:
+   :undoc-members:
