+++ docs/source/installation.rst
@@ -1,0 +1,17 @@
+Installation
+======
+.. code-block:: bash
+
+   pip install nsEVDx  
+   # Or clone from GitHub:
+   git clone https://github.com/Nischalcs50/nsEVDx
+   cd nsEVDx
+   pip install .
+
+For developers/contributors:
+
+.. code-block:: bash
+
+   git clone https://github.com/Nischalcs50/nsEVDx
+   cd nsEVDx
+   pip install -e .[dev]
