+++ docs/source/overview.rst
@@ -1,0 +1,75 @@
+Overview
+---------
+
+The **nsEVDx** library provides comprehensive tools for modeling extreme value distributions under both stationary and non-stationary conditions. Designed for hydrologists, climate scientists, and engineers working with extreme events, it implements both frequentist and Bayesian inference frameworks with flexible covariate modeling capabilities.
+
+Core Features
+-------------
+
+.. grid:: 1 2 3 3
+    :gutter: 3
+
+    .. grid-item-card:: GEV & GPD Models
+        :shadow: md
+
+        Complete support for Generalized Extreme Value and Generalized Pareto distributions.
+
+    .. grid-item-card:: Non-stationary Modeling
+        :shadow: md
+
+        Time-varying parameters with arbitrary covariates for changing climate conditions.
+
+    .. grid-item-card:: Bayesian & Frequentist
+        :shadow: md
+
+        Dual inference frameworks featuring robust MCMC sampling and MLE estimation.
+
+    .. grid-item-card:: Advanced MCMC
+        :shadow: md
+
+        High-performance MALA, HMC, and Metropolis-Hastings samplers for complex models.
+
+    .. grid-item-card:: Model Diagnostics
+        :shadow: md
+
+        Comprehensive convergence tracking and goodness-of-fit metrics for reliable results.
+
+    .. grid-item-card:: Minimal Dependencies
+        :shadow: md
+
+        Built on the core scientific stack: ``numpy``, ``scipy``, ``matplotlib``, and ``seaborn``.
+
+
+Key Applications
+-----------------
+
+.. grid:: 1 2 2 2
+    :gutter: 3
+
+    .. grid-item-card:: Climate Extremes
+        :text-align: left
+        :padding: 3
+
+        Analysis of extreme precipitation, temperature, and drought events 
+        under changing climate conditions.
+
+    .. grid-item-card:: Flood Frequency Analysis
+        :text-align: left
+        :padding: 3
+
+        Non-stationary flood frequency modeling with time-varying design 
+        quantiles for infrastructure planning.
+
+    .. grid-item-card:: Risk Assessment
+        :text-align: left
+        :padding: 3
+
+        Financial and environmental risk modeling with time-dependent 
+        extreme value parameters.
+
+    .. grid-item-card:: Engineering Design
+        :text-align: left
+        :padding: 3
+
+        Design value estimation for structures under non-stationary 
+        extreme loads.
