+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+nsEVDx
+======
+
+.. toctree::
+   :maxdepth: 4
+
+   nsEVDx
