+++ docs/source/index.rst
@@ -1,0 +1,19 @@
+.. nsEVDx documentation master file, created by
+   sphinx-quickstart on Mon Mar  9 15:49:28 2026.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+nsEVDx documentation
+====================
+
+
+.. toctree::
+   :maxdepth: 6
+   :caption: Contents:
+
+   overview
+   theory
+   installation
+   quickstart
+   modules
+     

[…]
