### 2026-03-09 404109b
added docs.yaml for publishing the web documentation
