### 2026-03-09 2aa1949
examples added to web docs and badge of web docs added to readme
