+++ docs/source/Examples1.rst
@@ -1,0 +1,10 @@
+Examples
+========
+
+You can view our example notebooks directly on GitHub:
+
+* `[Bayesian inference of Non-stationary GEV parameters] <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/example_GEV.ipynb>`_
+* `[Frequentist estimation of non-stationary GPD parameter and likelihood ratio test] <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/example_GPD_frequentist.ipynb>`_
+* `[Generating random variates from non-stationary GEV] <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/example_generating_rv_from_nsEVD.ipynb>`_
+* `[Bayesian metrics example] <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/example_bayesian_metrics.ipynb>`
+* `[Quickstart example] <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/Quick_start_example.ipynb>`
