+++ docs/source/index.rst
@@ -16,4 +16,5 @@ nsEVDx documentation
    installation
    quickstart
    modules
+   Examples1
      
