### 2026-03-10 442d6e7
error in Example1.rst rectified
