+++ docs/source/index.rst
@@ -1,7 +1,6 @@
 .. nsEVDx documentation master file, created by
    sphinx-quickstart on Mon Mar  9 15:49:28 2026.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+   
 
 nsEVDx documentation
 ====================
