### 2026-03-10 de60cfb
error in theory.rst
