### 2026-03-10 2e8d296
added citations and refs
