+++ docs/source/theory.rst
@@ -61,7 +61,7 @@ In a non-stationary framework, parameters are modeled as functions of covariates
 Where :math:`Z(t)` is a dynamic covariate that changes with time and affects the extreme value distributions.
 
 
-Non-Stationarity Configuration via Config vector
+Non-Stationarity Configuration via Config Vector
 -------------------------------------------------
 
 In ``nsEVDx``, non-stationarity is controlled via a configuration vector:
