+++ docs/source/citations_and_refs.rst
@@ -1,0 +1,32 @@
+Citations and References
+========================
+
+Citation
+--------
+
+If you use **nsEVDx** in your research, please cite the following:
+
+* Kafle, N., & Meier, C. I. (2025). *nsEVDx: A Python library for modeling Non-Stationary Extreme Value Distributions*. arXiv preprint `arXiv:2509.07261 <https://arxiv.org/abs/2509.07261>`_.
+* Kafle, N., & Meier, C. (2025). *nsEVDx: A Python Library for Modeling Non-Stationary Extreme Value Distributions (v0.1.0)*. Zenodo. `https://doi.org/10.5281/zenodo.15850043 <https://doi.org/10.5281/zenodo.15850043>`_
+
+References
+----------
+
+.. hlist::
+   :columns: 1
+
+   * Betancourt, M. (2017). *A Conceptual Introduction to Hamiltonian Monte Carlo*. arXiv: Methodology.
+   * Coles, S. (2007). *An introduction to statistical modeling of extreme values* (4th. printing). Springer. `https://doi.org/10.1007/978-1-4471-3675-0 <https://doi.org/10.1007/978-1-4471-3675-0>`_
+   * Gilleland, E. (2025). *extRemes: Extreme Value Analysis*. `https://doi.org/10.32614/CRAN.package.extRemes <https://doi.org/10.32614/CRAN.package.extRemes>`_
+   * Heffernan J. E., Stephenson A.G., & Gilleland E. (2003). *Ismev: An Introduction to Statistical Modeling of Extreme Values*. `https://CRAN.R-project.org/package=ismev <https://CRAN.R-project.org/package=ismev>`_
+   * Hosking, J. R. M., & Wallis, J. R. (1997). *Regional Frequency Analysis: An Approach Based on L-Moments* (Vol. 93). Cambridge University Press. `https://doi.org/10.1017/cbo9780511529443 <https://doi.org/10.1017/cbo9780511529443>`_
+   * IRSN. (2024). *NSGEV: Non-Stationary GEV Time Series*. `https://github.com/IRSN/NSGEV/ <https://github.com/IRSN/NSGEV/>`_
+   * Kafle, N., & Meier, C. (n.d.). *Detecting trends in short duration extreme precipitation over SEUS using neighborhood based method*. Manuscript in Preparation.
+   * Kafle, N., & Meier, C. (2025). *Evaluating Methodologies for Detecting Trends in Short-Duration Extreme Rainfall in the Southeastern United States*. Extreme Hydrological or Critical Event Analysis-III, EWRI Congress 2025, Anchorage, AK, U.S. `https://alaska2025.eventscribe.net <https://alaska2025.eventscribe.net>`_
+   * Kafle, N., Peleg, N., & Meier, C. I. *Detecting Spatially Consistent Trends in Sub-Hourly Extreme Rainfall Using a Neighborhood-Based Method*. AGU25.
+   * Oriol Abril-Pla et al. (2023). *PyMC: A modern, and comprehensive probabilistic programming framework in Python*. PeerJ Computer Science, 9, e1516. `https://doi.org/10.7717/peerj-cs.1516 <https://doi.org/10.7717/peerj-cs.1516>`_
+   * Paciorek, C. (2016). *climextRemes: Tools for Analyzing Climate Extremes*. `https://CRAN.R-project.org/package=climextRemes <https://CRAN.R-project.org/package=climextRemes>`_
+   * Robert, C. P., & Casella, G. (2009). *Introducing Monte Carlo Methods with R*. `https://doi.org/10.1007/978-1-4419-1576-4 <https://doi.org/10.1007/978-1-4419-1576-4>`_
[…]
