### 2026-03-10 22f6e51
added utility examples and command for auto generating .rst files for python modules in docs.yml
