+++ .github/workflows/docs.yml
@@ -30,6 +30,7 @@ jobs:
         
     - name: Build Sphinx docs
       run: |
+        sphinx-apidoc -o docs/source/ nsEVDx/ -f -e
         cd docs
         make html
 
