### 2026-03-10 ba88deb
added another example for L-moments and GPD para and enhanced docstrings
