+++ .github/workflows/docs.yml
@@ -30,7 +30,7 @@ jobs:
         
     - name: Build Sphinx docs
       run: |
-        sphinx-apidoc -o docs/source/ nsEVDx/ -f -e
+        # sphinx-apidoc -o docs/source/ nsEVDx/ -f -e # dont use this, this delays
         cd docs
         make html
 
