+++ docs/source/Examples1.rst
@@ -8,4 +8,4 @@ Please view our example notebooks directly on GitHub:
 * `Generating random variates from non-stationary GEV <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/example_generating_rv_from_nsEVD.ipynb>`_
 * `Bayesian metrics example <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/example_bayesian_metrics.ipynb>`_
 * `Quickstart example <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/Quick_start_example.ipynb>`_
-* `L-moments and GPD parameters example <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/L-moments_&_GPD_pars.ipynb>`_
+* `L-moments and GPD parameters example <https://github.com/Nischalcs50/nsEVDx/tree/master/examples/L-moments_and_GPD_pars.ipynb>`_
