### 2026-03-20 #265 @Nischalcs50
Thank you Aditya, for the detailed feedback!

I have pushed some initial updates addressing Akrita's comments and have integrated a Sphinx-based API for the documentation.

I am currently balancing the final stages of my dissertation and a few pending manuscripts, so I plan to dive back into the remaining items as soon as I have a clear window. I expect to have everything wrapped up and ready for another look within the next month.

Thanks for your patience and for helping improve the library!

Best regards,
Nischal
