+++ nsEVDx/__init__.py
@@ -1,16 +1,16 @@
 from .evd_model import NonStationaryEVD
 from .utils import (
+    EVD_parsViaMLE,
+    GEV_parsViaLM,
+    GPD_parsViaLM,
+    bayesian_metrics,
+    comb,
+    gelman_rubin,
+    l_moments,
     neg_log_likelihood,
     neg_log_likelihood_ns,
-    EVD_parsViaMLE,
-    comb,
-    l_moments,
-    GPD_parsViaLM,
-    GEV_parsViaLM,
-    plot_trace,
     plot_posterior,
-    bayesian_metrics,
-    gelman_rubin
+    plot_trace,
 )
 
 __all__ = [
@@ -25,10 +25,11 @@ __all__ = [
     "plot_trace",
     "plot_posterior",
     "bayesian_metrics",
-    "gelman_rubin"
+    "gelman_rubin",
 ]
 
 
 from ._version import get_versions
+
 __version__ = get_versions()["version"]
 del get_versions
