### 2026-03-26 5f8389b
issues with inf and -inf, invalid regions in neg ll and invalid priors, print to logging etc solved
