+++ .github/workflows/test.yml
@@ -5,23 +5,28 @@ on: [push, pull_request]
 jobs:
   build:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
 
     steps:
     - name: Check out repository
       uses: actions/checkout@v4
 
-    - name: Set up Python
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v5
       with:
-        python-version: 3.9
+        python-version: ${{ matrix.python-version }}
 
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
         pip install -e .[dev]
-        pip install pytest pytest-cov
 
+    - name: Run Ruff Check
+      run: ruff check .   
+      
     - name: Run Tests with coverage using pytest and codecov
       run: |
         pytest --cov=nsEVDx --cov-report=xml
@@ -30,3 +35,7 @@ jobs:
       uses: codecov/codecov-action@v5 
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
+        slug: Nischalcs50/nsEVDx
+      env:
+        OS: ubuntu-latest
+        PYTHON: ${{ matrix.python-version }}
