### 2026-03-27 70f4c6b
multi chain added to randwalk and MALA
