+++ nsEVDx/evd_model.py
@@ -1,4 +1,7 @@
+from __future__ import annotations
+
 import logging
+import warnings
 from typing import List, Tuple, Union
 
 import numpy as np
@@ -9,11 +12,19 @@ from scipy.stats import (
     rv_continuous,
     uniform,
 )
+from tqdm import tqdm
+
+try:
+    from joblib import Parallel, delayed
+    _JOBLIB_AVAILABLE = True
+except ImportError:
+    _JOBLIB_AVAILABLE = False
 
 from .utils import (
     GEV_parsViaLM,
     GPD_parsViaLM,
     neg_log_likelihood_ns,
+    gelman_rubin
 )
 
 logging.basicConfig(
@@ -23,7 +34,6 @@ logging.basicConfig(
     format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
 )
 logger = logging.getLogger("nsEVDx")
-
 
 # Supported distributions
 '''
@@ -37,6 +47,18 @@ SUPPORTED_DISTRIBUTIONS: Dict[str, rv_continuous] = {
     "gpd":      genpareto,     # Generalised Pareto Distribution
     }
 '''
+
+# Helper: acceptance rate warning
+def _check_acceptance(rate: float, sampler_name: str) -> None:
+    lo, hi = 0.20, 0.70
+    if not (lo <= rate <= hi):
+        warnings.warn(
+            f"[{sampler_name}] Acceptance rate {rate:.1%} is outside the "
+            f"recommended range [{lo:.0%}, {hi:.0%}]. "
+            "Consider tuning step_size / proposal_widths.",
+            stacklevel=3,
+        )
+
 # nsEVDx main model code
 class NonStationaryEVD:
     def __init__(self, config, data, cov, dist, prior_specs=None, bounds=None):
@@ -259,7 +281,7 @@ class NonStationaryEVD:
 
         return bounds
 
-    def log_prior(self, params):
+    def _log_prior(self, params):
         """
         Compute the log prior probability of the parameter vector.
 
@@ -329,7 +351,7 @@ class NonStationaryEVD:
 
         return logp
 
-    def neg_log_likelihood(self, params):
+    def _neg_log_likelihood(self, params):
         """
         Compute the negative log-likelihood for the given parameter vector.
 
@@ -352,7 +374,7 @@ class NonStationaryEVD:
             params, self.data, self.cov, self.config, self.dist
         )
 
-    def posterior_log_prob(self, params):
+    def _posterior_log_prob(self, params):
         """
         Compute the log posterior probability for the given parameter vector.
 
@@ -371,9 +393,9 @@ class NonStationaryEVD:
             The log posterior probability. If the prior is improper or
             evaluates to -inf, the result will reflect that.
         """
-        return -1 * self.neg_log_likelihood(params) + self.log_prior(params)
+        return -1 * self._neg_log_likelihood(params) + self._log_prior(params)
 
-    def numerical_grad_log_posterior(self, params, h=1e-2):
+    def _numerical_grad_log_posterior(self, params, h=1e-2):
         """
         Compute the numerical gradient of the log-posterior with respect to
         parameters.
@@ -404,19 +426,21 @@ class NonStationaryEVD:
         for i in range(len(params)):
             step = np.zeros_like(params)
             step[i] = h_vec[i]
-            f_plus = self.posterior_log_prob(params + step)
-            f_minus = self.posterior_log_prob(params - step)
+            f_plus = self._posterior_log_prob(params + step)
+            f_minus = self._posterior_log_prob(params - step)
             grad[i] = (f_plus - f_minus) / (2 * h_vec[i])
 
         return grad
 
-    def MH_Mala(
+    def _Mala_1chain(
         self,
         num_samples: int,
[…]
