### 2026-03-27 086b0f6
feat: multi chain for hmc added, fix:quick start and examples, feat: changelog and security added
