+++ SECURITY.md
@@ -1,0 +1,11 @@
+# Security Policy
+
+## Supported Versions
+| Version | Supported |
+| ------- | --------- |
+| 0.1.1   | ✅✅✅✅|
+
+## Reporting a Numerical Instability
+If you find a bug affecting EVD estimates, please open a GitHub Issue 
+or email nkafle@memphis.edu | nkafle.29@gmail.com.
+
