+++ nsEVDx/utils.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import List, Union
 
 import matplotlib.pyplot as plt
