+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+### Feat
+
+- multi chain for hmc added, fix:quick start and examples, feat: changelog and security added
+
 ## v0.1.1 (2025-09-29)
 
 ## v0.1.0 (2025-07-09)
