+++ nsEVDx/hmc_engine.py
@@ -1,0 +1,407 @@
+from __future__ import annotations
+
+import warnings
+from typing import Dict, List, Optional, Tuple, Union
+
+import numpy as np
+from tqdm.auto import tqdm
+
+from .utils import _total_log_prior, _grad_total_log_prior, gelman_rubin, neg_log_likelihood_ns
+from .utils import _total_log_prior, _grad_nll_gev, _grad_nll_gpd, _grad_total_log_prior
+
+try:
+    from joblib import Parallel, delayed
+    _JOBLIB_AVAILABLE = True
+except ImportError:
+    _JOBLIB_AVAILABLE = False
+    
+_LOG_SQRT_2PI = 0.9189385332046727   # log(sqrt(2π))
+
+class HMCEngine:
+    """
+    Analytical-gradient HMC / NUTS engine for NonStationaryEVD.
+
+    Parameters
+    ----------
+    model : NonStationaryEVD
+        A fitted (or partially set-up) model instance.  Must have
+        ``data``, ``cov``, ``config``, ``dist``, and ``prior_specs``
+        set.  Call ``model.prior_specs = model.suggest_priors()`` first
+        if you have not already done so.
+    grad_method : str
+        ``'analytical'`` (default) or ``'numerical'``.
+        Analytical is ~20-50x faster per leapfrog step.
+        Numerical is used automatically as fallback for unsupported dists.
+    """
+
+    _DELTA_MAX = 1000.0       # divergence threshold on ΔH
+
+    def __init__(self, model, grad_method: str = "analytical"):
+        self.model       = model
+        self.grad_method = grad_method
+        if isinstance(model.dist, str):
+            # If the user passed a string like 'genextreme'
+            dist = model.dist.lower()
+        else:
+            # If the user passed a SciPy object, get its internal name attribute
+            dist = getattr(model.dist, "name", "").lower()
+        self._is_gev = any(x in dist for x in ["genextreme", "gev"])
+        self._is_gpd = any(x in dist for x in ["genpareto", "gpd"])
+        if grad_method == "analytical" and not (self._is_gev or self._is_gpd):
+            warnings.warn(
+                f"Analytical gradients not implemented for '{dist}'. "
+                "Falling back to numerical gradients.",
+                stacklevel=2,
+            )
+            self.grad_method = "numerical"   
+ 
+    def _grad_log_posterior(self, params: np.ndarray) -> np.ndarray:
+        """
+        Calls the NonStationaryEVD analytical or numerical gradient logic.
+        """
+        return self.model._grad_log_posterior(params)
+
+    def _kinetic(self, momentum: np.ndarray, M_diag: np.ndarray) -> float:
+        """K(p) = 0.5 * p^T * M^-1 * p"""
+        with np.errstate(over='ignore'):
+            k = 0.5 * np.sum(momentum**2 / M_diag)
+        return k
+
+    def _hamiltonian(self, params: np.ndarray,
+                     momentum: np.ndarray,
+                     M_diag: np.ndarray,
+                     T: float = 1.0
+                     ) -> float:
+        """
+        Compute the Hamiltonian (total energy) of the system for HMC sampling.
+        Total Energy H(q,p) = -log_post(q) + K(p)
+        The Hamiltonian is the sum of the potential energy and kinetic energy.
+
+        In this context:
+
+        - Potential energy is defined as the negative log-posterior
+          (scaled by T), which encourages high-probability regions of
+          parameter space.
+        - Kinetic energy is computed as 0.5 * sum(momentum^2), assuming a
+          standard Gaussian momentum distribution.
+
+        Parameters
+        ----------
+        params : np.ndarray
+            Current position in parameter space (model parameters).
+        momentum : np.ndarray
[…]
