+++ nsEVDx/evd_model.py
@@ -21,6 +21,7 @@ except ImportError:
     _JOBLIB_AVAILABLE = False
 
 from .utils import GEV_parsViaLM, GPD_parsViaLM, gelman_rubin, neg_log_likelihood_ns
+from .utils import _total_log_prior, _grad_nll_gev, _grad_nll_gpd, _grad_total_log_prior
 
 logging.basicConfig(
     filename='nsEVDx_run.log',
@@ -105,6 +106,8 @@ class NonStationaryEVD:
         self.n_cov = cov.shape[0] if cov.ndim > 1 else 1
         self.prior_specs = prior_specs
         self.bounds = bounds
+        self._is_gev = dist in ("genextreme", "gev")
+        self._is_gpd = dist in ("genpareto",  "gpd")
         assert self.data.shape[0] == self.cov.shape[1], (
             "Mismatch between number of samples in data and covariates"
         )
@@ -214,7 +217,7 @@ class NonStationaryEVD:
 
         # Shape
         if self.config[2] == 0:
-            prior_specs.append(("normal", {"loc": 0, "scale": 0.1}))
+            prior_specs.append(("normal", {"loc": 0, "scale": 0.3}))
         else:
             # intercept
             prior_specs.append(("normal", {"loc": 0, "scale": 0.2}))
@@ -311,41 +314,24 @@ class NonStationaryEVD:
         if self.prior_specs is None:
             return 0.0
 
-        logp = 0.0
-        param_idx = 0
-        cov_counts = self.config  # [n_loc_cov, n_scale_cov, n_shape_cov]
-
-        for group in range(3):  # location, scale, shape
-            count = cov_counts[group]
-            n_params = count + 1 if count > 0 else 1  # intercept + covariates
-            # OR just stationary param
-            for _ in range(n_params):
-                ptype, kwargs = self.prior_specs[param_idx]
-                val = params[param_idx]
-
-                # Compute log-prior value
-                if ptype == "normal":
-                    prior_val = norm.logpdf(val, **kwargs)
-                elif ptype == "uniform":
-                    prior_val = uniform.logpdf(val, **kwargs)
-                elif ptype == "halfnormal":
-                    prior_val = halfnorm.logpdf(val, **kwargs)
-                else:
-                    raise ValueError(f"Unsupported prior type: {ptype}")
-
-                if not np.isfinite(prior_val):
-                    logger.debug(
-                                "Prior %s returned invalid logpdf for param"
-                                    "%d: val=%s, kwargs=%s",
-                                ptype, param_idx, val, kwargs
-                    )
-                    return -np.inf
-
-                logp += prior_val
-                param_idx += 1
-
-        return logp
-
+        return _total_log_prior(params, self.prior_specs)
+    
+    
+    def _grad_log_prior(self, params: np.ndarray) -> np.ndarray:
+        """
+        Compute gradient of log prior probability: 
+        ∂log π(θ)/∂θ 
+        
+        Parameters:
+            params : 1D array of current parameter values.
+            
+        Returns:
+            1D array of gradients for each prior.
+        """
+        # Pass the current parameters and the model's specific prior specs
+        return _grad_total_log_prior(params, self.prior_specs)
+    
+    
     def _neg_log_likelihood(self, params):
         """
         Compute the negative log-likelihood for the given parameter vector.
@@ -388,9 +374,52 @@ class NonStationaryEVD:
             The log posterior probability. If the prior is improper or
             evaluates to -inf, the result will reflect that.
         """
-        return -1 * self._neg_log_likelihood(params) + self._log_prior(params)
+        with np.errstate(over='ignore', invalid='ignore'):
+            x = -1 * self._neg_log_likelihood(params) + self._log_prior(params)
+        return x
+    
+    
[…]
