+++ nsEVDx/__init__.py
@@ -4,7 +4,11 @@ from .utils import (
     GEV_parsViaLM,
     GPD_parsViaLM,
     bayesian_metrics,
-    comb,
+    _comb,
+    _total_log_prior,
+    _grad_nll_gev,
+    _grad_nll_gpd, 
+    _grad_total_log_prior,
     gelman_rubin,
     l_moments,
     neg_log_likelihood,
@@ -12,9 +16,11 @@ from .utils import (
     plot_posterior,
     plot_trace,
 )
+from .hmc_engine import HMCEngine
 
 __all__ = [
     "NonStationaryEVD",
+    "HMCEngine",
     "neg_log_likelihood",
     "neg_log_likelihood_ns",
     "EVD_parsViaMLE",
