+++ tests/test_utils.py
@@ -1,54 +1,37 @@
 # test_utils.py
 import matplotlib.pyplot as plt
 import numpy as np
-from scipy.stats import genextreme
+from scipy.stats import genextreme, genpareto
 
 from nsEVDx.utils import (
-    EVD_parsViaMLE,
-    GEV_parsViaLM,
-    GPD_parsViaLM,
-    bayesian_metrics,
-    comb,
-    l_moments,
     neg_log_likelihood,
     neg_log_likelihood_ns,
-    plot_posterior,
+    _grad_nll_gev,
+    _grad_nll_gpd,
+    _total_log_prior,
+    _grad_total_log_prior,
+    EVD_parsViaMLE,
+    l_moments,
+    GPD_parsViaLM,
+    GEV_parsViaLM,
+    _comb,
+    _build_param_names,
     plot_trace,
+    plot_posterior,
+    bayesian_metrics,
+    gelman_rubin
 )
 
-
+#---- Basic Math & L-Moments ---
 def test_comb():
-    assert comb(5, 2) == 10
-
+    assert _comb(5, 2) == 10
+    assert _comb(5, 0) == 1
+    assert _comb(5, 6) == 0
 
 def test_l_moments():
     data = np.random.rand(100)
     l_moms = l_moments(data)
     assert len(l_moms) == 6
-
-
-def test_EVD_parsViaMLE():
-    data = np.random.rand(100)
-    params = EVD_parsViaMLE(data, genextreme)
-    assert len(params) == 3
-
-
-def test_neg_log_likelihood():
-    data = np.random.rand(100)
-    params = genextreme.fit(data)
-    nll = neg_log_likelihood(params, data, genextreme)
-    assert isinstance(nll, float)
-
-
-def test_neg_log_likelihood_ns():
-    dist = genextreme
-    data = dist.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
-    cov = np.vstack([np.ones_like(data), np.linspace(0, 1, len(data))])
-    config = [1, 0, 0]  # Time-varying location, constant scale and shape
-    params = [20, 0.02, 6.5, -0.2]
-    nll = neg_log_likelihood_ns(params, data, cov, config, dist)
-    assert np.isfinite(nll)
-
 
 def test_GPD_parsViaLM():
     data = np.random.rand(100)
@@ -60,21 +43,96 @@ def test_GEV_parsViaLM():
     data = np.random.rand(100)
     pars = GEV_parsViaLM(data)
     assert len(pars) == 3
+    
+def test_EVD_parsViaMLE():
+    data = np.random.rand(100)
+    params = EVD_parsViaMLE(data, genextreme)
+    assert len(params) == 3
 
+def test_neg_log_likelihood():
+    data = np.random.rand(100)
+    params = genextreme.fit(data)
+    nll = neg_log_likelihood(params, data, genextreme)
+    assert isinstance(nll, float)
 
+#---- Likelihoods ---
+def test_neg_log_likelihood_stationary():
+    dist = genextreme
+    data = dist.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
+    params = [20, 9, -0.12] # loc, scale, shape
+    nll = neg_log_likelihood(params, data, genextreme)
+    assert isinstance(nll, float)
+    assert nll > 0
+    
+def test_neg_log_likelihood_nonstationary():
+    dist = genextreme
+    data = dist.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
+    cov = np.vstack([np.ones_like(data), np.linspace(0, 1, len(data))])
+    config = [1, 0, 0]  # Time-varying location, constant scale and shape
+    params = [20, 0.02, 6.5, -0.2]
[…]
