+++ tests/test_evd_model.py
@@ -1,38 +1,338 @@
 # Tests for evd_model
 import numpy as np
-from scipy.stats import genextreme
-
+from scipy.stats import genextreme, genpareto
+import pytest
 from nsEVDx import NonStationaryEVD
+import warnings
+from nsEVDx.evd_model import _check_acceptance
 
-# Dummy inputs for testing
-np.random.seed(0)
-data = np.random.gumbel(loc=10, scale=5, size=50)
-cov = np.vstack([np.ones_like(data), np.linspace(0, 1, len(data))])
-config = [1, 0, 0]  # Time-varying location, constant scale and shape
-dist = genextreme
+#----_check_acceptance warnings
+def test_manual_warning_capture():
+    # Setup the capture bucket
+    with warnings.catch_warnings(record=True) as w:
+        warnings.simplefilter("always")
+        _check_acceptance(0.05, "TestSampler")
+        assert len(w) == 1
+        assert issubclass(w[-1].category, UserWarning)
+        assert "is outside the" in str(w[-1].message)
+        
+#---- Dummy inputs for testing
+RNG = np.random.default_rng(42)
+def _make_data(n=50, loc=20.0, scale=5.0, shape=-0.1, seed=0):
+    """Generate reproducible GEV data."""
+    np.random.seed(seed)
+    return genextreme.rvs(c=shape, loc=loc, scale=scale, size=n)
 
-"""
-prior_specs = [('normal',
-  {'loc': np.float64(8.7),
-   'scale': np.float64(4.35)}),
- ('normal', {'loc': 0, 'scale': 0.025}),
- ('normal', {'loc': np.float64(5.6), 'scale': 1}),
- ('normal', {'loc': 0, 'scale': 0.2})]
-"""
+def _make_cov_1d(n=50):
+    """1-covariate matrix: row of time indices normalised to [0,1]."""
+    return np.atleast_2d(np.linspace(0, 1, n))
 
-initial_params = [0.0, 0, 1.0, 0.1]
+def _make_cov_2d(n=50):
+    """2-covariate matrix."""
+    t = np.linspace(0, 1, n)
+    return np.vstack([t, t**2])
+
+def _model(config, n=50, dist=genextreme, prior_specs=None, bounds=None):
+    """Convenience: return a NonStationaryEVD with sensible data."""
+    data = _make_data(n=n)
+    n_cov = max(max(config), 1)
+    if n_cov == 1:
+        cov = _make_cov_1d(n)
+    else:
+        cov = _make_cov_2d(n)
+    return NonStationaryEVD(config, data, cov, dist,
+                            prior_specs=prior_specs, bounds=bounds)
+
+#----Testing NonStationaryEVD.__init__
+class TestInit:
+    def test_basic_construction_stationary(self):
+        m = _model([0, 0, 0])
+        assert isinstance(m, NonStationaryEVD)
+        assert m.data.shape[0] == 50
+        assert m.config == [0, 0, 0]
+        assert m.prior_specs is None
+        assert m.bounds is None
+    
+    def test_basic_construction_ns_loc(self):
+        m = _model([1, 0, 0])
+        assert m.n_cov >= 1
+    
+    def test_basic_construction_fully_ns(self):
+        m = _model([2, 1, 1])
+        # 2+1+1+3 = 7 parameters
+        assert len(m.descriptions) == 7
+    
+    def test_1d_cov_n_cov_branch(self):
+        """When cov is 1-D, n_cov should be 1."""
+        data = _make_data(n=30)
+        cov_1d = np.linspace(0, 1, 30)   # shape (30,) — not 2D
+        m = NonStationaryEVD([0, 0, 0], data, cov_1d, genextreme)
+        assert m.n_cov == 1
+    
+    def test_data_cov_shape_mismatch_raises(self):
[…]

[…]
