### 2026-04-01 6dd0c1d
fix: test_utils.py 93% complete and doing other tests
