+++ tests/test_utils.py
@@ -1,7 +1,8 @@
 # test_utils.py
 import matplotlib.pyplot as plt
 import numpy as np
-from scipy.stats import genextreme, genpareto
+from scipy.stats import genextreme, genpareto, chi2
+import pytest
 
 from nsEVDx.utils import (
     neg_log_likelihood,
@@ -48,6 +49,17 @@ def test_EVD_parsViaMLE():
     data = np.random.rand(100)
     params = EVD_parsViaMLE(data, genextreme)
     assert len(params) == 3
+    
+def test_EVD_parsViaMLE2():
+    data = genpareto.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
+    params = EVD_parsViaMLE(data, genpareto)
+    assert len(params) == 3
+    
+def test_EVD_parsViaMLE3():
+    data = genpareto.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
+    with pytest.raises(ValueError, match="Unsupported distribution"):
+        params = EVD_parsViaMLE(data, chi2)
+    
 
 def test_neg_log_likelihood():
     data = np.random.rand(100)
@@ -64,6 +76,22 @@ def test_neg_log_likelihood_stationary():
     assert isinstance(nll, float)
     assert nll > 0
     
+def test_neg_log_likelihood_stationary_exception():
+    class FailingDist:
+        def pdf(self, *args, **kwargs):
+            raise RuntimeError("forced failure")
+    data = np.array([1.0, 2.0, 3.0])
+    params = [0.0, 1.0, 0.1]
+    nll = neg_log_likelihood(params, data, FailingDist())
+    assert nll > 1e20
+    
+def test_nll_invalid_scale():
+    dist = genextreme
+    data = dist.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
+    params = [20, -1, -0.12] # loc, scale, shape
+    nll = neg_log_likelihood(params, data, genextreme)
+    assert nll > 1e20
+    
 def test_neg_log_likelihood_nonstationary():
     dist = genextreme
     data = dist.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
@@ -72,6 +100,67 @@ def test_neg_log_likelihood_nonstationary():
     params = [20, 0.02, 6.5, -0.2]
     nll = neg_log_likelihood_ns(params, data, cov, config, dist)
     assert np.isfinite(nll)
+    
+def test_neg_log_likelihood_full_nonstationary():
+    dist = genextreme
+    data = dist.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
+    cov = np.vstack([np.ones_like(data), np.linspace(0, 1, len(data))])
+    config = [1, 1, 1]  # Time-varying location, constant scale and shape
+    params = [20,0.01, 2,0.02,-0.1, 0.001]
+    nll = neg_log_likelihood_ns(params, data, cov, config, dist)
+    assert np.isfinite(nll)
+    
+def test_neg_log_likelihood_full_nonstationary2():
+    dist = genpareto
+    data = dist.rvs(-0.1, loc=22, scale=8, size=25, random_state=0)
+    cov = np.vstack([np.ones_like(data), np.linspace(0, 1, len(data))])
+    config = [1, 1, 1]  # Time-varying location, constant scale and shape
+    params = [20,0.01, 2,0.02,-0.1, 0.001]
+    nll = neg_log_likelihood_ns(params, data, cov, config, dist)
+    assert np.isfinite(nll)
+
+def test_neg_log_likelihood_bad_distribution():
+    data = np.random.randn(100)
+    cov = np.random.randn(100)
[…]
