+++ tests/test_evd_model.py
@@ -5,6 +5,12 @@ import pytest
 from nsEVDx import NonStationaryEVD
 import warnings
 from nsEVDx.evd_model import _check_acceptance
+from unittest.mock import MagicMock, patch
+try:
+    from joblib import Parallel, delayed
+    _JOBLIB_AVAILABLE = True
+except ImportError:
+    _JOBLIB_AVAILABLE = False
 
 #----_check_acceptance warnings
 def test_manual_warning_capture():
@@ -18,7 +24,10 @@ def test_manual_warning_capture():
         
 #---- Dummy inputs for testing
 RNG = np.random.default_rng(42)
-def _make_data(n=50, loc=20.0, scale=5.0, shape=-0.1, seed=0):
+
+
+def _make_data(dist=genextreme, n=50, 
+               loc=20.0, scale=5.0, shape=-0.1, seed=0):
     """Generate reproducible GEV data."""
     np.random.seed(seed)
     return genextreme.rvs(c=shape, loc=loc, scale=scale, size=n)
@@ -35,6 +44,17 @@ def _make_cov_2d(n=50):
 def _model(config, n=50, dist=genextreme, prior_specs=None, bounds=None):
     """Convenience: return a NonStationaryEVD with sensible data."""
     data = _make_data(n=n)
+    n_cov = max(max(config), 1)
+    if n_cov == 1:
+        cov = _make_cov_1d(n)
+    else:
+        cov = _make_cov_2d(n)
+    return NonStationaryEVD(config, data, cov, dist,
+                            prior_specs=prior_specs, bounds=bounds)
+
+def _modelgpd(config, n=50, dist=genpareto, prior_specs=None, bounds=None):
+    """Convenience: return a NonStationaryEVD with sensible data."""
+    data = _make_data(dist, n=n)
     n_cov = max(max(config), 1)
     if n_cov == 1:
         cov = _make_cov_1d(n)
@@ -150,7 +170,6 @@ class TestGetParamDescription:
 
 #----Testing suggest_priors()
 class TestSuggestPriors:
-
     def test_stationary_returns_3(self):
         m = _model([0, 0, 0])
         ps = m.suggest_priors()
@@ -189,7 +208,6 @@ class TestSuggestPriors:
 
 #----Testing suggest_bounds()
 class TestSuggestBounds:
-
     def test_gev_stationary(self):
         m = _model([0, 0, 0], dist=genextreme)
         bnds = m.suggest_bounds()
@@ -276,9 +294,9 @@ class TestLogPrior:
         val = m._log_prior(np.zeros(6))
         assert np.isfinite(val) 
 
-#---- Testing _neg_log_likelihood()
+#---- Testing _neg_log_likelihood(), _posterior_log_prob
+# _numerical_grad_log_posterior, _grad_log_posterior
 class TestNegLogLikelihood:
-
     def test_finite_for_valid_params(self):
         m = _model([1, 0, 0])
         params = np.array([20.0, 0.5, 5.0, -0.1])
@@ -292,11 +310,18 @@ class TestNegLogLikelihood:
         assert np.isfinite(val)
 
     def test_posterior_without_priors_equals_neg_nll(self):
-        """With flat prior (prior_specs=None), posterior = -NLL."""
         m = _model([0, 0, 0])
         params = np.array([20.0, 5.0, -0.1])
         assert m._posterior_log_prob(params) == pytest.approx(
             -m._neg_log_likelihood(params))
+        
+    def test_analytical_grad(self):
+        m = _model([1, 0, 0])
+        m.prior_specs = m.suggest_priors()
+        params = np.array([20.0, 0.5, 5.0, -0.1])
+        grad = m._grad_log_posterior(params)
+        assert grad.shape == params.shape
[…]
