### 2026-04-01 ee7b79a
fix: test for evd_model.py and utils.py >95%, othor minor errors and exceptions fixed
