+++ tests/test_evd_model.py
@@ -6,6 +6,7 @@ from nsEVDx import NonStationaryEVD
 import warnings
 from nsEVDx.evd_model import _check_acceptance
 from unittest.mock import MagicMock, patch
+from scipy.optimize import OptimizeResult
 try:
     from joblib import Parallel, delayed
     _JOBLIB_AVAILABLE = True
@@ -352,55 +353,129 @@ class TestNegLogLikelihood:
         m.prior_specs = m.suggest_priors()
         grad = m._grad_log_posterior(initial_params)
         assert grad.shape[0] == 4  
-
+    
+#----Testing gradients
+class TestGradients:
     def test_gradient_nan_fallback_to_zero(self):
-        """E
+        """
         Ensure gradient returns zeros if parameters 
         are outside support (v <= 0).
         """
         m = _model([1, 0, 0])
         bad_params = np.array([10.0, 100, 1.0, 5.0]) # mu, sigma, xi
         grad = m._grad_log_posterior(bad_params)
-        assert np.all(grad == 0)                      
+        assert np.all(grad == 0) 
+        
+    def test_gev_analytical_path(self):
+        """Tests the GEV analytical gradient 
+        branch in _grad_log_posterior."""
+        data = np.random.gumbel(loc=20, scale=5, size=30)
+        cov = np.atleast_2d(np.linspace(0, 1, 30))
+        config = [1, 0, 0] # B0, B1, sigma, xi
+        model = NonStationaryEVD(config, data, cov, "gev")
+        params = np.array([20.0, 0.1, 5.0, 0.1])
+        grad = model._grad_log_posterior(params)
+        assert grad.shape == (4,)
+        assert np.all(np.isfinite(grad))
+
+    def test_gpd_analytical_path(self):
+        """Tests the GPD analytical gradient branch 
+        in _grad_log_posterior."""
+        data = np.random.exponential(scale=10, size=30)
+        cov = np.atleast_2d(np.linspace(0, 1, 30))
+        config = [0, 0, 0] # mu, sigma, xi
+        model = NonStationaryEVD(config, data, cov, "gpd")
+        params = np.array([0.0, 10.0, 0.1])
+        grad = model._grad_log_posterior(params)
+        assert grad.shape == (3,)
+        assert np.all(np.isfinite(grad))
+
+    def test_grad_log_posterior_nan_to_zero(self):
+        """Verifies that NaN gradients (outside support) return zeros."""
+        data = np.array([50.0])
+        cov = np.atleast_2d([1.0])
+        model = NonStationaryEVD([0, 0, 0], data, cov, "gev")
+        # Force v = 1 - xi*(z) <= 0 by using a very large xi
+        bad_params = np.array([10.0, 1.0, 10.0])
+        # This should trigger the NaN-check 'if np.any(np.isnan(grad_nll))'
+        grad = model._grad_log_posterior(bad_params)
+        assert np.all(grad == 0.0) 
+        
+    def test_numerical_grad_log_posterior(self):
+        m = _model([1,0,0])
+        params = np.array([0, 100, 1.0, 100])
+        h = 1e-4
+        # Manually calculate one component to verify the return line logic
+        # f(x+h) - f(x-h) / (2*h)
+        f_plus = -1 * m._neg_log_likelihood(params + np.array([h, 0, 0, 0]))
+        f_minus = -1 * m._neg_log_likelihood(params - np.array([h, 0, 0, 0]))
+        expected_grad_0 = (f_plus - f_minus) / (2 * h)
+        grad = m._numerical_grad_log_posterior(params, h=h)
+        # Assert the returned array matches the expected manual calculation
+        assert np.allclose(grad[0], expected_grad_0)                   
         
 
 #---- Testing MCMC Methods
 class Test_mcmc_methods:
[…]
