### 2026-04-01 438df1d
fix: all tests and docstrings for hmc_engine
