+++ nsEVDx/utils.py
@@ -1,4 +1,3 @@
-import warnings
 from typing import List, Union
 
 import matplotlib.pyplot as plt
@@ -6,7 +5,6 @@ import numpy as np
 import seaborn as sns
 from scipy.optimize import minimize
 from scipy.special import gamma
-from scipy.stats import rv_continuous
 
 
 # Utility functions
@@ -115,15 +113,15 @@ def neg_log_likelihood_ns(
     # Ensure parameters are valid
     if np.any(sigma <= 0):
         return Safe_INF
-    
+
     # Standardized variates
     z = (data - mu) / sigma
-    
+
     dist_name = dist.name.lower() if hasattr(dist, 'name') else str(dist).lower()
     if dist_name in ["gev", "genextreme"]:
         v = 1.0 - xi * z
         if np.any(v <= 0): return Safe_INF
-        
+
         # Gumbel Limit Handling (xi -> 0)
         eps = 1e-7
         log_v = np.log(v)
@@ -133,11 +131,11 @@ def neg_log_likelihood_ns(
             -np.log(sigma) - z - np.exp(-z),  # Gumbel/Extreme Type I
             -np.log(sigma) + (1.0/xi_safe - 1.0) * log_v - v**(1.0/xi_safe)
         )
-        
+
     elif dist_name in ["gpd", "genpareto"]:
         w = 1.0 + xi * z
         if np.any(w <= 0): return Safe_INF
-        
+
         eps = 1e-7
         log_w = np.log(w)
         xi_safe = np.where(np.abs(xi) < eps, eps, xi)
@@ -195,7 +193,7 @@ def _grad_nll_gev(
     else:
         sigma = np.full(n, params[idx]); idx_sig = idx; idx += 1
 
-    # shape 
+    # shape
     if config[2] >= 1:
         nc = int(config[2])
         K  = params[idx: idx + nc + 1]
@@ -205,7 +203,7 @@ def _grad_nll_gev(
         xi = np.full(n, params[idx]); idx_xi = idx
 
     z = (data - mu) / sigma
-    v = 1.0 - xi * z                  
+    v = 1.0 - xi * z
     if np.any(v <= 0):
         return np.full(len(params), np.nan)
 
@@ -214,7 +212,7 @@ def _grad_nll_gev(
     xi_safe  = np.where(np.abs(xi) < eps, np.sign(xi + 1e-30) * eps, xi)
     small_xi = np.abs(xi) < eps
     inv_v = 1.0 / v
-    
+
     # Partial derivatives of LL
     # ---dlogL/dv--------------
     dL_dv = np.where(
@@ -223,15 +221,15 @@ def _grad_nll_gev(
         (1.0/xi_safe - 1.0)*inv_v - (1.0/xi_safe)*v**(1.0/xi_safe - 1.0)
     )
 
-    #---- Location gradient (dv/dmu = xi/sigma) 
+    #---- Location gradient (dv/dmu = xi/sigma)
     dL_dmu = dL_dv * (xi / sigma)
-    
+
     #---- Scale gradient (dv/dsigma = xi*z/sigma)
     # dL_dsig_base = -1.0/sigma + dL_dv * (xi * z / sigma)
     # because of exponential regression for the scale
     # chain rule: d(log σ) = dsigma/sigma
-    dL_da = -1.0 + dL_dv * (xi * z)  
-    
+    dL_da = -1.0 + dL_dv * (xi * z)
+
     #---- Shape gradient (dv/dxi = -z)
     log_v  = np.log(v)
     dL_dxi = np.where(
@@ -242,7 +240,7 @@ def _grad_nll_gev(
         + v**(1.0/xi_safe)*log_v/xi_safe**2
[…]
