+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - multi chain for hmc added, fix:quick start and examples, feat: changelog and security added
 
+### Fix
+
+- all tests and docstrings for hmc_engine
+- test for evd_model.py and utils.py >95%, othor minor errors and exceptions fixed
+- test_utils.py 93% complete and doing other tests
+- some tests added
+
 ## v0.1.1 (2025-09-29)
 
 ## v0.1.0 (2025-07-09)
