+++ setup.cfg
@@ -19,6 +19,7 @@ install_requires =
     scipy
     matplotlib
     seaborn
+    tqdm
 
 [options.packages.find]
 where = nsEVDx

[…]
