### 2026-04-01 78c653d
fix: all but examples
