+++ docs/source/conf.py
@@ -35,11 +35,10 @@ exclude_patterns = []
 
 
 # -- Options for HTML output -------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
-
 # html_theme = 'alabaster'
 html_theme = "sphinx_rtd_theme"
 html_static_path = ["_static"]
 
 # Optional: control docstring format
 autodoc_typehints = "description"
+add_module_names = False
