### 2026-04-01 69a1268
fix: examples and quick start fixed
