### 2026-04-01 c03e4db
fix:paper md documentation reference
