### 2026-04-01 4997130
feat: release.yml file added to automate pipy release
