### 2026-04-01 #265 @Nischalcs50
Hi @akritaag and @adityapt,

Thank you for your thorough reviews and suggestions!

I have made several updates to the package addressing the majority of the review comments. Could you please take another look and let me know if everything is now in good shape? I am also planning to submit this package to JOSS.

Key updates include:

-Expanded test coverage, especially for edge cases (likelihoods, priors, gradients, and failure paths)
-Most of the computations are numpy-based, which increased the speed of computation
-Fixed bugs in tests (including duplicate test naming and missing cases)
-Added CI workflow with multiple Python versions and coverage reporting
-Code cleaned with ruff, PEP8 issues resolved
-Automated release by adding release.yml workflow
-Improved packaging consistency (aligned dependencies in pyproject.toml)
-Cleaned up code style (formatting, consistency, minor fixes)
-Updated README with badges and clearer setup/usage instructions
-Fixed issues in CONTRIBUTING and added missing repo files (e.g., CHANGELOG)
-Added a new HMC engine class with support for multiple chains to improve convergence behavior and diagnostics

Please let me know if you see any remaining gaps or areas that need tightening.

Best regards,
Nischal Kafle
