### 2026-04-02 77d20a9
fix:fixed error in initializing NonStationaryEVD object due to shape
