### 2026-04-02 17f2987
Fix: error in intitializing nsEVD object and changelog
