+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ### Feat
 
+- release.yml file added to automate pipy release
 - multi chain for hmc added, fix:quick start and examples, feat: changelog and security added
 
 ### Fix
 
+- gitignore updated
+- final
+- examples and quick start fixed
+- all but examples
 - all tests and docstrings for hmc_engine
 - test for evd_model.py and utils.py >95%, othor minor errors and exceptions fixed
 - test_utils.py 93% complete and doing other tests
