### 2026-07-04 456c74f
Revise README with PyOpenscience badges
Updated README to include additional badges and improve formatting.
