### 2026-07-05 35fc08d
Bump version to 0.2.0
