### 2026-07-05 bf1de25
Bump version to 0.2.2
