### 2026-07-05 0e9dc07
Bump version to v0.2.3 and update release to PiPy trusted
