+++ .github/workflows/release.yml
@@ -3,33 +3,37 @@ name: Release to PyPI
 on:
   push:
     tags:
-      - 'v*'  # Trigger workflow only on tags starting with "v", e.g., v1.2.3
+      - "v*"
+
+permissions:
+  contents: read
+  id-token: write
 
 jobs:
   build-and-publish:
     runs-on: ubuntu-latest
+
     steps:
       - name: Checkout code
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
-      - name: Set up Python 
-        uses: actions/setup-python@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
         with:
           python-version: "3.12"
-          cache: 'pip'
+          cache: "pip"
 
-      - name: Install build tools
+      - name: Install build dependencies
         run: |
-          python -m pip install --upgrade pip setuptools wheel build twine
+          python -m pip install --upgrade pip
+          python -m pip install build
 
-      - name: Build package
-        run: python -m build --sdist --wheel
+      - name: Build distributions
+        run: |
+          python -m build
 
-      - name: Publish to PyPI
-        env:
-          TWINE_USERNAME: __token__  # Use PyPI API token
-          TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
-        run: python -m twine upload dist/*
+      - name: Publish package to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
 
       - name: Confirm release
-        run: echo "Published tag $GITHUB_REF to PyPI successfully!"
+        run: echo "Successfully published $GITHUB_REF to PyPI."
