### 2026-07-05 d7fa431
updated zenedo version
Updated library version and citation details in README.
