### 2026-07-06 e504274
PyOpenSci issue no updated
