+++ Paper/paper.md
@@ -10,7 +10,7 @@ authors:
 affiliations:
 - name: Department of Civil Engineering, University of Memphis, USA
   index: 1
-date: 2025-07-08
+date: 2026-07-05
 software_repository_url: https://github.com/Nischalcs50/nsEVDx
 archive_url: https://zenodo.org/record/XXXXXXX
 license: MIT
@@ -28,17 +28,14 @@ csl: apa.csl
 
 # Summary
 
-`nsEVDx` is an open-source Python package for fitting stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. It can be used to model extreme events in fields like hydrology, climate science, finance, and insurance, using both frequentist and Bayesian methods. For Bayesian inference it employs advanced Monte Carlo sampling techniques such as the Metropolis-Hastings, Metropolis-adjusted Langevin (MALA), and Hamiltonian Monte Carlo (HMC) algorithms. Unlike many existing extreme value theory (EVT) tools, which can be complex or lack Bayesian options, `nsEVDx` offers an intuitive, python-native interface that is both user-friendly and extensible. It requires only standard scientific python libraries (`numpy`, `scipy`) for its core functionality, while optional features like plotting and diagnostics use `matplotlib` and `seaborn`. A key feature of `nsEVDx` is its flexible support for non-stationary modeling, where the location, scale, and shape parameters can each depend on arbitrary, user-defined covariates. This enables practical applications such as linking extremes to other variables (e.g., rainfall extremes to temperature or maximum stock market losses to market volatility indices). Overall, `nsEVDx` aims to serve as a practical, easy-to-use, and extensible tool for researchers and practitioners analyzing extreme events in non-stationary environments.
+`nsEVDx` is an open-source Python package for fitting stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. It can be used to model extreme events in fields like hydrology, climate science, finance, and insurance, using both frequentist and Bayesian methods. For Bayesian inference it employs advanced Monte Carlo sampling techniques such as the Metropolis-Hastings, Metropolis-adjusted Langevin (MALA), and Hamiltonian Monte Carlo (HMC) algorithms. Existing EVD softwares often emphasize either frequentist inference or require users to formulate models within general probabilistic programming frameworks, leaving relatively few Python packages that provide a dedicated interface for both stationary and non-stationary EVD modeling with built-in Bayesian inference. To bridge this gap, `nsEVDx` offers an intuitive, Python-native interface that is both user-friendly and extensible. It requires only standard scientific python libraries (`numpy`, `scipy`) for its core functionality, while optional features like plotting and diagnostics use `matplotlib` and `seaborn`. A key feature of `nsEVDx` is its flexible support for non-stationary modeling, where the location, scale, and shape parameters can each depend on arbitrary, user-defined covariates. This enables practical applications such as linking extremes to other variables (e.g., rainfall extremes to temperature or maximum stock market losses to market volatility indices). Overall, `nsEVDx` aims to serve as a practical, easy-to-use, and extensible tool for researchers and practitioners analyzing extreme events in non-stationary environments.
 
 # Statement of Need
 
 Probabilistic modeling of extreme events is essential across disciplines, from resilient infrastructure design and climate adaptation to insurance pricing and financial risk management. In many real-world processes, the statistical properties of the extremes are often non-stationary, driven by long-term changes such as climate change, urbanization, or economic shifts. Accurately estimating return periods and risks under these evolving conditions requires fitting non-stationary extreme value distributions (EVDs) to observations.
[…]
