### 2026-07-06 fd735d6
fix: changelog and paper updated
