+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+## v0.2.3 (2026-07-04)
+
+## v0.2.2 (2026-07-04)
+
+## v0.2.1 (2026-07-04)
+
+## v0.2.0 (2026-07-04)
+
 ### Feat
 
 - release.yml file added to automate pipy release
