### 2026-07-09 98a6054
Bump version to v0.2.4
