### 2026-07-17 acc4218
changes in the paper.md
