+++ Paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'nsEVDx: A Python library for modeling non-stationary extreme value distributions'
+title: 'nsEVDx: A Python library for stationary and non-stationary extreme value modeling'
 authors:
 - name: Nischal Kafle
   orcid: 0009-0004-3187-4920
@@ -25,6 +25,15 @@ tags:
 bibliography: paper.bib
 csl: apa.csl
 ---
+# Summary
+
+Infrastructure design, insurance pricing, and financial planning all rely  on understanding the probability of rare, extreme events, e.g.,extreme floods, rainfall events, and stock market crashes. Traditionally, researchers calculated the probability of extreme event of a given magnitude by assuming that their chance of occurrence remain stationary over time. However, due to shifting global conditions like climate change, rapid urbanization, and evolving economy, probability of these extreme events are changing, they have become "non-stationary." 
+
+`nsEVDx` is an open-source Python package for fitting stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data.
+
+`nsEVDx` is an open-source Python tool built to help researchers and engineers accurately calculate these shifting risks. It provides a user-friendly, reliable framework that allows users to link the statistical properties of extreme events to changing real-world factors (such as connecting extreme rainfall patterns directly to rising temperatures or linking stock market losses to shifts in market volatility). 
+
+The software packages advanced statistical modeling techniques into an accessible format, offering both traditional calculations and advanced predictive simulation methods. By handling the complex mathematics behind the scenes, `nsEVDx` allows scientists, engineers, and risk analysts to easily build more accurate safety models, leading to more resilient infrastructure designs and better-informed climate adaptation strategies.
 
 # Summary
 
@@ -32,10 +41,30 @@ csl: apa.csl
 
 # Statement of Need
 
+Traditional extreme value frequency analysis assumes that the statistical properties of extreme events remain stationary over time. However, driven by environmental and socioeconomic shifts such as climate change, rapid urbanization, and structural market volatility, real-world extremes are increasingly non-stationary. Accurately quantifying shifting risks and return periods requires statistical tools that can seamlessly incorporate arbitrary, time-varying covariates into all parameters of extreme value distributions (EVDs).
+
+`nsEVDx` is designed to solve the critical accessibility gap between highly flexible but complex general-purpose probabilistic programming languages (PPLs) and rigid, domain-specific packages that lack native Python execution. Building custom, mathematically sound, non-stationary EVD models with custom prior boundaries and gradient-based Bayesian samplers requires significant programmatic and statistical expertise, which presents a steep barrier to entry for applied researchers. 
+
+The target audience for `nsEVDx` consists of hydrologists, climate scientists, infrastructure engineers, and financial risk analysts who require a transparent, reliable, and mathematically rigorous pipeline for non-stationary extreme value modeling. By lowering the programmatic overhead needed to implement both frequentist optimization and advanced Markov Chain Monte Carlo (MCMC) inference algorithms (e.g., MALA and HMC), the software allows domain experts to focus on interpreting non-stationary trends and making defensible engineering decisions under evolving environmental conditions.
+
+
+# Statement of the Field
+
[…]
