### 2026-07-28 b3733db
added Software Design in paper.md
