+++ Paper/paper.md
@@ -26,22 +26,13 @@ bibliography: paper.bib
 csl: apa.csl
 ---
 # Summary
+Infrastructure design, insurance pricing, and financial planning all rely on understanding the probability of rare, extreme events, such as extreme floods, rainfall events, and stock market crashes. Traditionally, researchers estimated the probability of extreme events of a given magnitude by assuming that their probability of occurrence remained stationary over time. However, due to changing global conditions such as climate change, urbanization, and economic shifts, the probability of extreme events has become non-stationary.
+`nsEVDx` is an open-source Python package developed to help researchers quantify these changing risks by fitting both stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. A key feature of `nsEVDx` is its flexible support for non-stationary modeling, where the location, scale, and shape parameters can each depend on user-defined covariates. This enables practical applications such as linking rainfall extremes to temperature or maximum stock market losses to market volatility indices. Overall, `nsEVDx` aims to serve as a practical, easy-to-use, and extensible tool for researchers and practitioners analyzing extreme events in non-stationary environments. The software packages complex statistical modeling techniques into an accessible interface, offering both traditional frequentist and advanced Bayesian methods.
 
-Infrastructure design, insurance pricing, and financial planning all rely  on understanding the probability of rare, extreme events, e.g.,extreme floods, rainfall events, and stock market crashes. Traditionally, researchers calculated the probability of extreme event of a given magnitude by assuming that their chance of occurrence remain stationary over time. However, due to shifting global conditions like climate change, rapid urbanization, and evolving economy, probability of these extreme events are changing, they have become "non-stationary." 
-
-`nsEVDx` is an open-source Python package for fitting stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data.
-
-`nsEVDx` is an open-source Python tool built to help researchers and engineers accurately calculate these shifting risks. It provides a user-friendly, reliable framework that allows users to link the statistical properties of extreme events to changing real-world factors (such as connecting extreme rainfall patterns directly to rising temperatures or linking stock market losses to shifts in market volatility). 
-
-The software packages advanced statistical modeling techniques into an accessible format, offering both traditional calculations and advanced predictive simulation methods. By handling the complex mathematics behind the scenes, `nsEVDx` allows scientists, engineers, and risk analysts to easily build more accurate safety models, leading to more resilient infrastructure designs and better-informed climate adaptation strategies.
-
-# Summary
-
[…]
