### 2026-07-28 35da732
Fix stationary-scale gradient off by a factor of sigma (GEV/GPD)
The analytic scale gradient applied the log-link chain-rule factor
(sigma * dNLL/dsigma) unconditionally. In the stationary branch
(config[1] == 0) the parameter is sigma itself, so _grad_nll_gev and
_grad_nll_gpd returned a scale gradient too large by a factor of sigma,
corrupting the momentum kick for sigma in the HMC/MALA samplers. Drop the
factor in the identity-link branch. Adds a finite-difference regression
test over GEV/GPD, stationary and log-link scale.
