+++ Paper/paper.bib
@@ -167,4 +167,12 @@ Pages: 2.2-1},
 	year = {2025},
 }
 
+@phdthesis{kafle2026raingauge,
+  author = {Kafle, Nischal},
+  title = {Rain-Gauge Network Effects on the Uncertainty and Trends in Short-Duration Extreme Precipitation},
+  school = {The University of Memphis},
+  year = {2026},
+  type = {PhD dissertation}
+  doi = {add_later_from_PRO_QUEST}
+}
 
