### 2026-07-28 49810ad
statement of field and refs
