+++ Paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'nsEVDx: A Python library for stationary and non-stationary extreme value modeling'
+title: 'nsEVDx: A Python Library for Stationary and Non-Stationary Extreme Value Modeling'
 authors:
 - name: Nischal Kafle
   orcid: 0009-0004-3187-4920
@@ -26,13 +26,12 @@ bibliography: paper.bib
 csl: apa.csl
 ---
 # Summary
-Infrastructure design, insurance pricing, and financial planning all rely on understanding the probability of rare, extreme events, such as extreme floods, rainfall events, and stock market crashes. Traditionally, researchers estimated the probability of extreme events of a given magnitude by assuming that their probability of occurrence remained stationary over time. However, due to changing global conditions such as climate change, urbanization, and economic shifts, the probability of extreme events has become non-stationary.
-`nsEVDx` is an open-source Python package developed to help researchers quantify these changing risks by fitting both stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. A key feature of `nsEVDx` is its flexible support for non-stationary modeling, where the location, scale, and shape parameters can each depend on user-defined covariates. This enables practical applications such as linking rainfall extremes to temperature or maximum stock market losses to market volatility indices. Overall, `nsEVDx` aims to serve as a practical, easy-to-use, and extensible tool for researchers and practitioners analyzing extreme events in non-stationary environments. The software packages complex statistical modeling techniques into an accessible interface, offering both traditional frequentist and advanced Bayesian methods.
+Infrastructure design, insurance pricing, and financial planning all rely on understanding the probability of rare, extreme events, such as extreme floods, rainfall events, and stock market crashes. Traditionally, researchers estimated the probability of extreme events of a given magnitude by assuming that their probability of occurrence remained stationary over time. However, changing global conditions such as climate change, urbanization, and economic shifts can cause the probability of extreme events to become non-stationary. `nsEVDx` is an open-source Python package developed to help researchers quantify these changing risks by fitting both stationary and non-stationary Extreme Value Distributions (EVDs) to extreme value data. It supports non-stationary modeling in which the location, scale, and shape parameters can each depend on user-defined covariates. This enables practical applications such as linking rainfall extremes to temperature or maximum stock market losses to market volatility indices. Overall, `nsEVDx` provides researchers and practitioners with an easy-to-use, flexible, and extensible Python-native interface for frequentist and Bayesian inference, including MCMC methods.
 
 
 # Statement of Need
 
-Traditional extreme value frequency analysis assumes that the statistical properties of extreme events remain stationary over time. However, driven by environmental and socioeconomic shifts such as climate change, rapid urbanization, and structural market volatility, the probability distribution of extremes are increasingly observed to exhibit non-stationary behavior in hydroclimatic systems [cite_something]. Accurately quantifying shifting risks and return periods requires statistical tools that can seamlessly incorporate arbitrary, time-varying covariates into all parameters of extreme value distributions (EVDs).
[…]
