+++ Paper/paper.bib
@@ -124,11 +124,12 @@ Pages: 2.2-1},
 	file = {PDF:C\:\\Users\\nkafle\\Zotero\\storage\\6KQHYWXA\\Gilleland - 2004 - extRemes Extreme Value Analysis.pdf:application/pdf},
 }
 
-@misc{irsn_nsgev_2024,
-	title = {{NSGEV}: {Non}-{Stationary} {GEV} {Time} {Series}},
-	url = {https://github.com/IRSN/NSGEV/},
-	author = {{IRSN}},
-	year = {2024},
+@Manual{irsn_nsgev_2024,
+	title = {NSGEV: Non-Stationary GEV Time Series},
+  	author = {Yves Deville},
+  	year = {2026},
+  	note = {R package version 0.2.3},
+  	url = {https://github.com/IRSN/NSGEV/},
 }
 
 @article{roberts_exponential_1996,
@@ -172,7 +173,95 @@ Pages: 2.2-1},
   title = {Rain-Gauge Network Effects on the Uncertainty and Trends in Short-Duration Extreme Precipitation},
   school = {The University of Memphis},
   year = {2026},
-  type = {PhD dissertation}
+  type = {PhD dissertation},
   doi = {add_later_from_PRO_QUEST}
 }
 
+@article{prosdocimi2015detection,
+  author = {Prosdocimi, Ilaria and Kjeldsen, Thomas R. and Miller, John D.},
+  title = {Detection and attribution of urbanization effect on flood extremes using nonstationary flood-frequency models},
+  journal = {Water Resources Research},
+  volume = {51},
+  number = {6},
+  pages = {4244--4262},
+  year = {2015},
+  doi = {https://doi.org/10.1002/2015WR017065}
+}
+
+@article{jayaweera2025evidence,
+  author = {Jayaweera, Lalani and Wasko, Conrad and Nathan, Rory},
+  title = {Evidence for Non-Stationarity in the GEV Shape Parameter When Modeling Extreme Rainfall},
+  journal = {Water Resources Research},
+  volume = {61},
+  number = {5},
+  pages = {e2023WR036426},
+  year = {2025},
+  doi = {10.1029/2023WR036426}
+}
+
+@article{eccles2025substantial,
+  author = {Eccles, Rohan and Syktus, Jozef and Trancoso, Ralph and Chapman, Sarah and Wasko, Conrad and Evans, Jason P. and Thatcher, Marcus and Di Virgilio, Giovanni and Stassen, Christian},
+  title = {Substantial increases in future precipitation extremes---insights from a large ensemble of downscaled CMIP6 models},
+  journal = {npj Natural Hazards},
+  volume = {2},
+  number = {1},
+  pages = {60},
+  year = {2025},
+  doi = {10.1038/s44304-025-00107-1}
+}
+
+@inproceedings{kafle2025detecting,
+  author = {Kafle, Nischal and Peleg, Nadav and Meier, Claudio I.},
+  title = {Detecting Spatially Consistent Trends in Sub-Hourly Extreme Rainfall Using a Neighborhood-Based Method},
+  booktitle = {AGU Fall Meeting Abstracts},
+  volume = {2025},
+  pages = {H13G-07},
+  year = {2025},
+  month = {December},
+  url = {https://ui.adsabs.harvard.edu/abs/2025AGUFMH13G...07K/abstract}
+}
+
+@article{katz2002statistics,
+  author = {Katz, Richard W. and Parlange, Marc B. and Naveau, Philippe},
+  title = {Statistics of Extremes in Hydrology},
+  journal = {Advances in Water Resources},
+  volume = {25},
+  number = {8--12},
+  pages = {1287--1304},
+  year = {2002},
[…]
