### 2026-07-29 15b9808
paper.md edits
