### 2026-07-29 11caec8
figures added to paper.md
