### 2026-07-29 208f63b
paper.md changed figure1
