+++ Paper/paper.md
@@ -49,27 +49,28 @@ These differences motivate a dedicated Python tool that combines flexibility wit
 
 
 # Software Design
-`nsEVDx` is designed to provide a transparent and researcher-focused framework for fitting stationary and non-stationary extreme value distributions while minimizing the programming burden typically associated with custom Bayesian implementations (Figure 1).
+`nsEVDx` is designed to provide a transparent and researcher-focused framework for fitting stationary and non-stationary extreme value distributions while minimizing the programming burden typically associated with custom Bayesian implementations. The overall workflow implemented in `nsEVDx` is shown in Figure 1.
 
-![Figure 1: Workflow implemented in `nsEVDx`. Extreme-value observations and covariates are translated into stationary or non-stationary GEV/GPD models through the configuration vector `config=[a,b,c]`. Models can be fit using either frequentist optimization or Bayesian MCMC sampling. Diagnostic metrics and visualizations are generated within the same workflow.](Fig1.png)
+![Workflow implemented in `nsEVDx` for specifying, fitting, and diagnosing stationary and non-stationary GEV and GPD models using frequentist optimization or Bayesian MCMC sampling](Fig1.png)
 
 The concepts of non-stationarity and MCMC techniques used in `nsEVDx` are based on the foundational texts by @christian_p_robert_introducing_2009 and @coles_introduction_2007. The core class `NonStationaryEVD` handles parameter parsing and specification, log-likelihood evaluation, prior assignment, optimization, and MCMC sampling. This unified design allows users to switch between frequentist and Bayesian workflows without redefining model structures. Frequentist methods uses `scipy.optimize` to minimize the non-stationary negative log-likelihood, while the Bayesian MCMC methods are implemented in numpy, allowing full transparency and customization. In Bayesian estimation, `nsEVDx` can infer prior specifications based on the data and configuration or accept user-defined priors. In the frequentist mode, it can determine suitable parameter bounds automatically. However, user-defined priors or bounds are recommended for better convergence and interpretability. The implementation of L-moments in some utility methods follows the approach described by @j_r_m_hosking_regional_1997. Currently, `nsEVDx` supports linear modeling for the location and shape parameters, and exponential (log-linear) modeling for the scale parameter, to ensure positivity.
 
 `nsEVDx` controls non-stationarity via a configuration vector `config = [a, b, c]`, where each entry specifies the number of covariates used to model the location, scale, and shape parameters of the EVD. Entry with a value of `0` implies stationarity (i.e., no covariate dependence), while integer values `> 0` indicate non-stationary modeling using the corresponding number of covariates for the parameter. This representation provides a compact and scalable mechanism for defining a wide range of stationary and non-stationary models while maintaining a consistent user interface (Figure 2). 
 
-![Figure 2. Configuration-vector framework used in `nsEVDx` for specifying stationary and non-stationary EVD models.](Fig2.png)
+![Configuration-vector framework used in `nsEVDx` for specifying stationary and non-stationary EVD models.](Fig2.png)
 
 Currently, location and shape parameters can be modeled using linear relationships, while the scale parameter is modeled using an exponential link function to ensure positivity. Polynomial relationships can be accommodated by transforming covariates before model fitting. Future releases aim to incorporate spline-based and mixed-population formulations.
 
-Another key design decision is the implementation of Bayesian samplers directly in `NumPy` rather than relying exclusively on external probabilistic programming frameworks. This approach maximizes transparency, enables users to inspect and modify algorithmic details, and reduces dependencies. The package currently provides Random-Walk Metropolis-Hastings, Metropolis-Adjusted Langevin Algorithm (MALA), and Hamiltonian Monte Carlo (HMC) samplers, allowing users to balance computational efficiency and inferential robustness according to their application. 
[…]
