### 2026-07-29 8fe664c
fix: sphinx documentation and refs
