+++ Paper/paper.bib
@@ -11,9 +11,7 @@
 	doi = {10.7717/peerj-cs.1516},
 	pmcid = {10495961},
 	pmid = {37705656},
-	note = {MAG ID: 4386373863
-S2ID: 27127b9617c4a66f2a4a89f7ca57695623b3c218},
-	pages = {e1516--e1516},
+	pages = {e1516--e1516}
 }
 
 @article{michael_betancourt_conceptual_2017,
@@ -23,10 +21,7 @@ S2ID: 27127b9617c4a66f2a4a89f7ca57695623b3c218},
 	author = {Betancourt, Michael},
 	month = jan,
 	year = {2017},
-	note = {ARXIV\_ID: 1701.02434
-MAG ID: 2572702041
-S2ID: 59b3ebd49ef620b3051d9777dfc59076d846949f},
-}
+	}
 
 @misc{heffernan_j_e_ismev_2003,
 	title = {ismev: {An} {Introduction} to {Statistical} {Modeling} of {Extreme} {Values}},
@@ -38,8 +33,7 @@ S2ID: 59b3ebd49ef620b3051d9777dfc59076d846949f},
 	month = apr,
 	year = {2003},
 	note = {Institution: Comprehensive R Archive Network
-Pages: 1.42},
-	file = {PDF:C\:\\Users\\nkafle\\Zotero\\storage\\7NNLMRJT\\Original S functions written by Janet E. Heffernan with R port and R  documentation provided by Alec G. Stephenson. - 2003 - ismev An Introduction to Statistical Modeling of Extreme Values.pdf:application/pdf},
+Pages: 1.42}
 }
 
 @misc{paciorek_climextremes_2016,
@@ -52,15 +46,14 @@ Pages: 1.42},
 	month = aug,
 	year = {2016},
 	note = {Institution: Comprehensive R Archive Network
-Pages: 0.3.1},
-	file = {PDF:C\:\\Users\\nkafle\\Zotero\\storage\\V84TFRB3\\Paciorek - 2016 - climextRemes Tools for Analyzing Climate Extremes.pdf:application/pdf},
+Pages: 0.3.1}
 }
 
 @misc{noauthor_pystan_2023,
 	title = {{PyStan}: {The} python interface to {Stan}},
 	url = {https://pystan.readthedocs.io/},
 	author = {Stan development Team,},
-	year = {2023},
+	year = {2023}
 }
 
 @misc{noauthor_cmdstan_2023,
@@ -78,9 +71,7 @@ Pages: 0.3.1},
 	month = apr,
 	year = {1997},
 	url = {https://ui.adsabs.harvard.edu/abs/1997rfa..book.....H},
-	publisher = {Cambridge University Press},
-	note = {MAG ID: 2056199992
-S2ID: dafa032edadcea535f84f60334af40a3ef1cef6c},
+	publisher = {Cambridge University Press}
 }
 
 @book{christian_p_robert_introducing_2009,
@@ -89,10 +80,8 @@ S2ID: dafa032edadcea535f84f60334af40a3ef1cef6c},
 	author = {Robert, Christian P. and Casella, George},
 	month = nov,
 	year = {2009},
-	doi = {10.1007/978-1-4419-1576-4},
-	note = {MAG ID: 2162188693
-S2ID: 1cf894b52919a330c3b477024b203d03058b34e3},
-}
+	doi = {10.1007/978-1-4419-1576-4}
+	}
 
 @book{coles_introduction_2007,
 	address = {London Berlin Heidelberg},
@@ -104,8 +93,7 @@ S2ID: 1cf894b52919a330c3b477024b203d03058b34e3},
 	language = {en},
 	publisher = {Springer},
 	author = {Coles, Stuart},
[…]
