### 2026-07-30 771c43e
refs doi added
