+++ Paper/paper.bib
@@ -6,7 +6,7 @@
 	abstract = {PyMC is a probabilistic programming library for Python that provides tools for constructing and fitting Bayesian models. It offers an intuitive, readable syntax that is close to the natural syntax statisticians use to describe models. PyMC leverages the symbolic computation library PyTensor, allowing it to be compiled into a variety of computational backends, such as C, JAX, and Numba, which in turn offer access to different computational architectures including CPU, GPU, and TPU. Being a general modeling framework, PyMC supports a variety of models including generalized hierarchical linear regression and classification, time series, ordinary differential equations (ODEs), and non-parametric models such as Gaussian processes (GPs). We demonstrate PyMC's versatility and ease of use with examples spanning a range of common statistical models. Additionally, we discuss the positive role of PyMC in the development of the open-source ecosystem for probabilistic programming.},
 	journal = {PeerJ Computer Science},
 	author = {{Oriol Abril-Pla} and {Virgile Andreani} and {C. Carroll} and {L. Y. Dong} and {Christopher Fonnesbeck} and {Maxim Kochurov} and {Ravin Kumar} and {Junpeng Lao} and {Christian C. Luhmann} and {Osvaldo A. Martin} and {Michael Osthege} and {Ricardo Vieira} and {Thomas V. Wiecki} and {Robert Zinkov}},
-	month = sep,
+	month = {sep},
 	year = {2023},
 	doi = {10.7717/peerj-cs.1516},
 	pmcid = {10495961},
@@ -19,8 +19,9 @@
 	abstract = {Hamiltonian Monte Carlo has proven a remarkable empirical success, but only recently have we begun to develop a rigorous understanding of why it performs so well on difficult problems and how it is best applied in practice. Unfortunately, that understanding is confined within the mathematics of differential geometry which has limited its dissemination, especially to the applied communities for which it is particularly important. In this review I provide a comprehensive conceptual account of these theoretical foundations, focusing on developing a principled intuition behind the method and its optimal implementations rather of any exhaustive rigor. Whether a practitioner or a statistician, the dedicated reader will acquire a solid grasp of how Hamiltonian Monte Carlo works, when it succeeds, and, perhaps most importantly, when it fails.},
 	journal = {arXiv: Methodology},
 	author = {Betancourt, Michael},
-	month = jan,
+	month = {jan},
 	year = {2017},
+	doi = {https://doi.org/10.48550/arXiv.1701.02434}
 	}
 
 @misc{heffernan_j_e_ismev_2003,
@@ -30,10 +31,11 @@
 	abstract = {Functions to support the computations carried out in `An Introduction to Statistical Modeling of Extreme Values' by Stuart Coles. The functions may be divided into the following  groups; maxima/minima, order statistics, peaks over thresholds and point processes.},
 	urldate = {2025-07-01},
 	author = {{Heffernan J. E.} and {Stephenson A.G.} and {Gilleland E.}},
-	month = apr,
+	month = {apr},
 	year = {2003},
 	note = {Institution: Comprehensive R Archive Network
-Pages: 1.42}
+Pages: 1.42},
+	doi = {10.32614/CRAN.package.ismev}
 }
 
 @misc{paciorek_climextremes_2016,
@@ -43,24 +45,27 @@ Pages: 1.42}
[…]
