### 2026-07-30 016bbca
dois corrected and figs update
