+++ Paper/paper.bib
@@ -21,7 +21,7 @@
 	author = {Betancourt, Michael},
 	month = {jan},
 	year = {2017},
-	doi = {https://doi.org/10.48550/arXiv.1701.02434}
+	doi = {10.48550/arXiv.1701.02434}
 	}
 
 @misc{heffernan_j_e_ismev_2003,
@@ -77,7 +77,7 @@ Pages: 0.3.1},
 	year = {1997},
 	url = {https://ui.adsabs.harvard.edu/abs/1997rfa..book.....H},
 	publisher = {Cambridge University Press},
-	doi = {https://doi.org/10.1017/CBO9780511529443}
+	doi = {10.1017/CBO9780511529443}
 }
 
 @book{christian_p_robert_introducing_2009,
@@ -95,7 +95,7 @@ Pages: 0.3.1},
 	series = {Springer series in statistics},
 	title = {An introduction to statistical modeling of extreme values},
 	isbn = {978-1-85233-459-8},
-	doi = {https://doi.org/10.1007/978-1-4471-3675-0},
+	doi = {10.1007/978-1-4471-3675-0},
 	language = {en},
 	publisher = {Springer},
 	author = {Coles, Stuart},
@@ -230,7 +230,7 @@ Pages: 2.2-1}
   title = {Extreme Value Theory in Engineering},
   publisher = {Academic Press},
   year = {1988},
-  doi = {https://doi.org/10.1016/C2009-0-22169-6}
+  doi = {10.1016/C2009-0-22169-6}
 }
 
 @article{phan2019composable,
@@ -238,5 +238,5 @@ Pages: 2.2-1}
     author  = {Phan, Du and Pradhan, Neeraj and Jankowiak, Martin},
     journal = {arXiv preprint arXiv:1912.11554},
     year    = {2019},
-	doi = {https://doi.org/10.48550/arXiv.1912.11554}
+	doi = {10.48550/arXiv.1912.11554}
 }
