### 2026-08-11 c946711
Removed deprecated API_docs
