+++ API_docs/API.md
@@ -1,678 +1,0 @@
-<a id="nsEVDx"></a>
-
-# nsEVDx
-
-<a id="nsEVDx.evd_model"></a>
-
-# nsEVDx.evd_model
-
-<a id="nsEVDx.evd_model.NonStationaryEVD"></a>
-
-## NonStationaryEVD Objects
-
-``` python
-class NonStationaryEVD()
-```
-
-<a id="nsEVDx.evd_model.NonStationaryEVD.__init__"></a>
-
-## \_\_init\_\_
-
-``` python
-def __init__(config, data, cov, dist, prior_specs=None, bounds=None)
-```
-
-Instantiate a object(reffered to as 'sampler') of class NonStationaryEVD.
-
-##### Parameters
-
-**config** : list of int. Represents a non-stationarity configuration for [location, scale, shape].
-
-**What is config ?**
-
-    Config denotes the configuration of the non-stationary model. In 
-    config = [p, q, r], p, q, and r denote the number of covariates for the location
-    , scale, and shape parameters of the EVD. A value of 0 implies stationarity;
-    a value >0 indicates non-stationarity with that many covariates.
-
-For examples:
-
-    config = [0,0,0] indicates [stationary_location, stationary_scale
-    ,stationary_shape], a full stationary configuration.
-
-    config = [1,0,0] indicates [locatin modeled with 1 covariate, stationary
-    scale, stationary shape].
-
-    config = [2,1,0] indicates [location modeled with 2 covariates, scale
-    modeled with 1 covariate, stationary shape].
-
-> *Note:* The location and shape parameters are modeled linearly, whereas the scale parameter is modeled exponentially:
-
-    - `location_model`: "constant" or "linear"
-    - `scale_model`: "constant" or "exponential"
-    - `shape_model`: "constant" or "linear"
-
-    Internally, these options apply a regression of the form:
-
-    -   Linear: θ(t) = θ₀ + θ₁·X(t)
-
-    -   Exponential: θ(t) = exp(θ₀ + θ₁·X(t)), where X represent
-        covariates
-
-**data** : array-like Observed extremes in chronlogical order (e.g., annual maxima).
-
-**cov** : array-like Covariate matrix, shape (n_covariates, n_samples).
-
-**dist** : scipy.stats distribution object (genextreme or genpareto).
-
-**prior_specs** : list of tuples Optional prior specifications for each parameter. Required if performing bayesian sampling.
-
-Format: `[(dist_name, params_dict), ...]`
-
-`e.g., [('normal', {'loc': 0, 'scale': 10}), ('uniform',{'loc': 0, 'scale': 5}), ...]`
-
-**bounds** : List of tuples. Optional bounds for each parameter, required if estimating the the parameters by frequentist approach.
-
-##### Returns
-
-**NonStationaryEVD** **object :** An instance of the NonStationaryEVD class initialized with the specified configuration, data, covariates, and distribution.
-
-<a id="nsEVDx.evd_model.NonStationaryEVD.get_param_description"></a>
-
-## get_param_description
-
-``` python
-@staticmethod
-def get_param_description(config: List[int], n_cov: int) -> List[str]
-```
-
-Returns a list of strings describing each parameter's role in the parameter vector, based on the provided configuration.
-
-##### Parameters
-
-**config** : list of int Non-stationarity configuration [location, scale, shape].
-
-**n_cov** : int Total number of covariates available.
-
-##### Returns
-
-**list of str**
-
-Descriptions of each parameter in order.
-
-For instance, config = [1,0,0] and n_cov = 1, would have four parameters as follows:
-
-`['B0 (location intercept)', 'B1 (location slope for covariate 1)', 'scale, 'shape']`
-
[…]
