### 2026-08-11 d85ed3a
removed unnecessary files
