### 2026-08-11 3072152
fix: corrected some bugs in hmc_engine and fixed README quick start example
