### 2026-08-12 a1bf744
PEP8 check; ruff pass
