### 2026-08-23 46da09d
Report MH acceptance rates over production samples only
_Mala_1chain and _RandWalk_1chain counted acceptances across all
num_samples + burn_in iterations but divided by num_samples, so any
nonzero burn-in inflated the reported rate (it could exceed 1.0) and
steered step-size tuning against the 20-70 percent guidance. Counting
now starts after burn-in, matching HMCEngine._sample_hmc.
