### 2026-08-23 0bb4929
Paper: rename heading for the State of the field check, fix typos, strip DOI prefix
- 'Statement of the Field' -> 'State of the field' so editorialbot's
  required-section check passes
- typos: effort effort, practicioners (x2), explictly, advance -> advanced,
  uses -> use
- prosdocimi2015detection: bare DOI (the https://doi.org/ prefix made the
  reference check flag it INVALID)
