+++ Paper/paper.bib
@@ -159,7 +159,7 @@ Pages: 2.2-1}
   number = {6},
   pages = {4244--4262},
   year = {2015},
-  doi = {https://doi.org/10.1002/2015WR017065}
+  doi = {10.1002/2015WR017065}
 }
 
 @article{jayaweera2025evidence,
