### 2026-08-24 3b46160
few changes in the paper
