+++ Paper/paper.bib
@@ -89,7 +89,7 @@ Pages: 0.3.1},
 	doi = {10.1007/978-1-4419-1576-4}
 	}
 
-@book{coles_introduction_2007,
+@book{coles_introduction_2001,
 	address = {London Berlin Heidelberg},
 	edition = {4th. printing},
 	series = {Springer series in statistics},
@@ -100,7 +100,7 @@ Pages: 0.3.1},
 	publisher = {Springer},
 	author = {Coles, Stuart},
 	month = {nov},
-	year = {2007}
+	year = {2001}
 }
 
 @misc{gilleland_extremes_2025,
@@ -148,7 +148,8 @@ Pages: 2.2-1}
   school = {The University of Memphis},
   year = {2026},
   type = {PhD dissertation},
-  doi = {ADD_LATER_FROM_PRO_QUEST}
+  number = {32785789},
+  url = {https://ezproxy.memphis.edu:3443/login?url=https://www.proquest.com/dissertations-theses/rain-gauge-network-effects-on-uncertainty-trends/docview/3369343212/se-2}
 }
 
 @article{prosdocimi2015detection,
@@ -222,7 +223,7 @@ Pages: 2.2-1}
   title = {(Forthcoming.) Robustness of Regionally Derived, Short-Duration Rainfall Depth-Duration-Frequency Estimates to the Choice of Minimum Interevent Time: Evidence across Climates, Raingauge Densities, and Regionalization Approaches},
   journal = {Journal of Hydrologic Engineering},
   year = {2026},
-  doi = {10.1061/JHYEFF/HEENG-6729}
+  doi = {10.1061/JHYEFF.HEENG-6729}
 }
 
 @book{castillo1988extreme,
