### 2026-08-25 87bae35
fix: mass matrix, tests and examples
