### 2026-08-25 5f1102b
fix: Improved GPD support check by rejecting z<0
