### 2026-08-25 1c7bf89
fix: addressing ISSUE #11: API contracts and drifted docstrings
