### 2026-08-26 01a9527
fixed a typo in README.md
