### 2026-08-26 0b5248b
Fix: Issue #19- pytest-cov added to requirement, now tests should run w/o this problem
