+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fix
 
+- addressing ISSUE #11: API contracts and drifted docstrings
 - Improved GPD support check by rejecting z<0
 - mass matrix, tests and examples
 - corrected some bugs in hmc_engine and fixed README quick start example
