### 2026-08-26 55ca15b
Fix: Issue 18: now tests dont open figure windows
