### 2026-08-27 77bf030
Fix: Please check if #18, #19 are addressed; Manuscript related comments from the reviewer @dosvk is partially addresses-work in progress....; Extended theoretical description in the documentation; added references
