+++ Paper/paper.bib
@@ -241,3 +241,29 @@ Pages: 2.2-1}
     year    = {2019},
 	doi = {10.48550/arXiv.1912.11554}
 }
+
+@article{emcee,
+   author = {{Foreman-Mackey}, D. and {Hogg}, D.~W. and {Lang}, D. and {Goodman}, J.},
+    title = {emcee: The MCMC Hammer},
+  journal = {PASP},
+     year = 2013,
+   volume = 125,
+    pages = {306-312},
+   eprint = {1202.3665},
+      doi = {10.1086/670067}
+}
+
+@incollection{neal2011mcmc,
+  author    = {Neal, Radford M.},
+  title     = {{MCMC} Using {H}amiltonian Dynamics},
+  booktitle = {Handbook of Markov Chain Monte Carlo},
+  editor    = {Brooks, Steve and Gelman, Andrew and Jones, Galin L. and Meng, Xiao-Li},
+  publisher = {CRC Press},
+  address   = {Boca Raton, FL},
+  pages     = {113--162},
+  year      = {2011},
+  doi       = {10.1201/b10905},
+  eprint    = {1206.1901},
+  archiveprefix = {arXiv},
+  primaryclass  = {stat.CO}
+}
