+++ docs/source/theory.rst
@@ -18,6 +18,48 @@ with location :math:`\mu`, scale :math:`\sigma > 0`, and shape :math:`\xi` is:
 
    F(x; \mu, \sigma, 0) = \exp\left\{-\exp\left[-\left(\frac{x - \mu}{\sigma}\right)\right]\right\} \quad \text{for } \xi = 0
 
+The probability density function (PDF) of GEV is given by:
+
+.. math::
+
+   f(x; \mu, \sigma, \xi)
+   =
+   \frac{1}{\sigma}
+   \left[
+      1 - \xi\left(\frac{x-\mu}{\sigma}\right)
+   \right]^{1/\xi - 1}
+   \exp\left\{
+      -
+      \left[
+         1 - \xi\left(\frac{x-\mu}{\sigma}\right)
+      \right]^{1/\xi}
+   \right\},
+   \qquad \xi \neq 0
+
+with support
+
+.. math::
+
+   1 - \xi\left(\frac{x-\mu}{\sigma}\right) > 0.
+
+For :math:`\xi = 0` (Gumbel case),
+
+.. math::
+
+   f(x; \mu, \sigma, 0)
+   =
+   \frac{1}{\sigma}
+   \exp\left[
+      -
+      \left(\frac{x-\mu}{\sigma}\right)
+   \right]
+   \exp\left\{
+      -
+      \exp\left[
+         -
+         \left(\frac{x-\mu}{\sigma}\right)
+      \right]
+   \right\}.
 
 Generalized Pareto Distribution
 -------------------------------
@@ -34,6 +76,35 @@ The CDF of exceedances :math:`Y = X - \mu > 0` over threshold :math:`\mu` follow
 
    F(y; \mu, \sigma, 0) = 1 - \exp\left(-\frac{y - \mu}{\sigma}\right) \quad \text{for } \xi = 0
 
+The PDF of GPD is given by:
+
+.. math::
+
+   f(y; \mu, \sigma, \xi)
+   =
+   \frac{1}{\sigma}
+   \left(
+      1 + \xi\frac{y-\mu}{\sigma}
+   \right)^{-1/\xi - 1},
+   \qquad \xi \neq 0
+
+with support
+
+.. math::
+
+   1 + \xi\frac{y-\mu}{\sigma} > 0.
+
+For :math:`\xi = 0` (Exponential case),
+
+.. math::
+
+   f(y; \mu, \sigma, 0)
+   =
+   \frac{1}{\sigma}
+   \exp\left(
+      -
+      \frac{y-\mu}{\sigma}
+   \right).
 
 Non-stationary Framework
 ------------------------
@@ -76,3 +147,114 @@ Each element in the configuration specifies the number of covariates for the **l
 * Values **> 0** indicate non-stationary modeling using the corresponding number of covariates.
 
 This framework allows flexible, parsimonious modeling of non-stationary extreme value distributions, including covariates only where supported by data.
+
+
+Log-Likelihood
+--------------
+
+The primary computational core in ``nsEVDx`` is the evaluation of the generalized
+log-likelihood  suitable for both stationary and non-stationary three-parameter
+extreme value distributions (EVDs).
+The log-likelihood form the foundation of parameter estimation, frequentist inference,
+and Bayesian inference for both stationary and non-stationary EVDs.
+
+For a sample of observations
+:math:`\{x_1, x_2, \ldots, x_n\}`, the log-likelihood is defined as
+
+.. math::
+
+   \ell(\theta)
+   =
+   \sum_{i=1}^{n}
+   \log f(x_i;\theta),
+
+where :math:`f(\cdot)` denotes the PDF of the selected EVD and
+:math:`\theta` represents the full parameter vector.
+
[…]
