### 2026-08-28 a474734
Fix: added figure in ms, changed prior specs and bounds to adapt differently for GEV and GPD; changed how divergence is computed in the HMC engine
