### 2026-08-28 2ddf73f
Feat: added truncated normal distribution for priors and updated grad of priors; in the earlier commit, I made changes to scale epsilon with parameter while initializing mass matrix
