+++ tests/test_evd_model.py
@@ -186,6 +186,14 @@ class TestSuggestPriors:
         ps = m.suggest_priors()
         assert len(ps) == 6
 
+    def test_shape_prior_is_truncated(self):
+        for dist in (genextreme, genpareto):
+            m = _model([0, 0, 0], dist=dist)
+            shape_prior = m.suggest_priors()[-1]
+            assert shape_prior[0] == "truncatednormal"
+            assert shape_prior[1]["low"] == -0.5
+            assert shape_prior[1]["high"] == 0.5
+
     def test_ns_loc_only(self):
         # [1,0,0] → 4 params: B0, B1, sigma, xi
         m = _model([1, 0, 0])
@@ -193,7 +201,7 @@ class TestSuggestPriors:
         assert len(ps) == 4
 
     def test_gpd_location_prior_respects_support(self):
-        data = genpareto.rvs(0.2, loc=0, scale=5, size=50,
+        data = genpareto.rvs(0.2, loc=5, scale=5, size=50,
                               random_state=0)
         m = NonStationaryEVD([0, 0, 0], data, _make_cov_1d(50), genpareto)
 
@@ -201,10 +209,8 @@ class TestSuggestPriors:
 
         assert location_prior[0] == "uniform"
         assert location_prior[1]["loc"] < np.min(data)
-        assert location_prior[1]["loc"] + location_prior[1]["scale"] == pytest.approx(
-            np.min(data)
-        )
-
+        assert location_prior[1]["loc"] + location_prior[1]["scale"] <= np.min(data)
+       
     def test_ns_scale_only(self):
         # [0,1,0] → 4 params: mu, a0, a1, xi
         m = _model([0, 1, 0])
